% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acv4elasticNet.R
\name{acv4enet_GLMNET_SEMCpp}
\alias{acv4enet_GLMNET_SEMCpp}
\title{acv4enet_GLMNET_SEMCpp}
\usage{
acv4enet_GLMNET_SEMCpp(
  SEM,
  subsets,
  raw,
  weights,
  lambda,
  alpha,
  hessianOfDifferentiablePart = NULL,
  control
)
}
\arguments{
\item{SEM}{model of class Rcpp_SEMCpp. Models of this class
can be generated with the SEMFromLavaan-function.}

\item{subsets}{list with subsets created with createSubsets()}

\item{raw}{controls if the internal transformations of lessSEM should be used.}

\item{weights}{labeled vector with weights of regularized parameters}

\item{lambda}{value of tuning parameter lambda}

\item{alpha}{value of tuning parameter alpha (for elastic net)}

\item{control}{arguments passed to GLMNET inner iteration}
}
\description{
internal function for approximate cross-validation based on the internal model representation of lessSEM. The GLMNET part refers to the fact
that this function uses the GLMNET optimizer for non-differentiable penalty functions.
}
