% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aCV4model.R
\name{aCV4lavaan}
\alias{aCV4lavaan}
\title{aCV4lavaan}
\usage{
aCV4lavaan(lavaanModel, k, raw = FALSE)
}
\arguments{
\item{lavaanModel}{model of class lavaan}

\item{k}{the number of cross-validation folds. We recommend leave-one-out cross-validation; i.e. set k to the number of persons in the data set}

\item{raw}{controls if the cross-validation should use the internal transformations of aCV4SEM. aCV4SEM will use an exponential function for all variances to 
avoid negative variances. This can result in better sub-group parameters, but may not be necessary and will also result in more difficult to interpret parameters.}
}
\description{
approximate cross-validation for models of class lavaan.
}
