% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aCV4model.R
\name{aCV4regsem}
\alias{aCV4regsem}
\title{aCV4regsem}
\usage{
aCV4regsem(regsemModel, lavaanModel, k, penalty, lambda, eps = 1e-04)
}
\arguments{
\item{regsemModel}{model of class regsem}

\item{lavaanModel}{model of class lavaan. This must be the same model used to set up the regsem model!}

\item{k}{the number of cross-validation folds. We recommend leave-one-out cross-validation; i.e. set k to the number of persons in the data set}

\item{penalty}{which penalty was used in regsem? Currently available are: penalty = "lasso", penalty = "ridge", or penalty = "elasticNet"}

\item{lambda}{value of the tuning parameter lambda. The tuning parameter alpha for the elastic net will be exported from the regsemModel}

\item{eps}{controls the smooth approximation of non-differential penalty functions (e.g., lasso, adaptive lasso, or elastic net). Smaller values result in closer approximation, but may also cause larger issues in optimization.}
}
\description{
approximate cross-validation for models of class regsem.
}
