% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aCV4model.R
\name{aCV4regularizedSEMWithCustomPenalty}
\alias{aCV4regularizedSEMWithCustomPenalty}
\title{aCV4regularizedSEMWithCustomPenalty}
\usage{
aCV4regularizedSEMWithCustomPenalty(regularizedSEMWithCustomPenalty, k)
}
\arguments{
\item{k}{the number of cross-validation folds. We recommend leave-one-out cross-validation; i.e. set k to the number of persons in the data set}

\item{regularizedSEMWithCustomPenalty.}{model of class regularizedSEMWithCustomPenalty.}
}
\description{
approximate cross-validation for models of class regularizedSEMWithCustomPenalty. These models can be fit with regularizedSEMWithCustomPenalty.() (see ?regularizedSEMWithCustomPenalty.)
in this package.
}
