% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acv4elasticNet.R
\name{acv4ridge}
\alias{acv4ridge}
\title{acv4ridge}
\usage{
acv4ridge(
  regularizedSEM,
  k,
  recomputeHessian = TRUE,
  returnSubsetParameters = FALSE,
  control = controlGlmnet()
)
}
\arguments{
\item{regularizedSEM}{model of class regularizedSEM}

\item{k}{the number of cross-validation folds. We recommend leave-one-out cross-validation; i.e. set k to the number of persons in the data set. Alternatively, 
a matrix with pre-defined subsets can be passed to the function. See ?linr::aCV4regularizedSEM for an example}

\item{recomputeHessian}{if set to FALSE, the Hessians from the quasi newton optimization with GLMNET will be used. Otherwise the Hessian will be recomputed. We currently recommend setting recomputeHessian to TRUE}

\item{returnSubsetParameters}{if set to TRUE, the parameter estimates of the individual cross-validation training sets will be returned}

\item{control}{parameters passed to the GLMNET optimizer. Note that only arguments of the inner iteration are used. See ?controlGlmnet for more details}
}
\description{
approximate cross-validation for ridge regularized SEM
}
