\name{aCV4SEM-package}
\alias{aCV4SEM-package}
\alias{aCV4SEM}
\docType{package}
\title{\packageTitle{aCV4SEM}}
\description{\packageDescription{aCV4SEM}}
\section{Package Content}{\packageIndices{aCV4SEM}}
\author{\packageAuthor{aCV4SEM}}
\section{Maintainer}{\packageMaintainer{aCV4SEM}}
\keyword{package}
