% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aI4model.R
\name{aI4regularizedSEM}
\alias{aI4regularizedSEM}
\title{aI4regularizedSEM}
\usage{
aI4regularizedSEM(regularizedSEM, k, recomputeHessian = TRUE)
}
\arguments{
\item{regularizedSEM}{model of class regularizedSEM}

\item{k}{the number of cross-validation folds. We recommend leave-one-out cross-validation; i.e. set k to the number of persons in the data set}

\item{recomputeHessian}{if set to false, the Hessians from the quasi newton optimization with GLMNET will be used. Otherwise the Hessian will be recomputed.}
}
\description{
approximate influence for models of class regularizedSEM. These models can be fit with regularizedSEM() (see ?regularizedSEM)
in this package.
}
