% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-influenceDetails.R
\name{AIC,approximateInfluence-method}
\alias{AIC,approximateInfluence-method}
\title{AIC}
\usage{
\S4method{AIC}{approximateInfluence}(object)
}
\arguments{
\item{object}{object of class approximateInfluence}
}
\description{
returns the AIC
}
