% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-gpRegularized.R
\name{AIC,gpRegularized-method}
\alias{AIC,gpRegularized-method}
\title{AIC}
\usage{
\S4method{AIC}{gpRegularized}(object)
}
\arguments{
\item{object}{object of class gpRegularized}
}
\description{
returns the AIC
}
