% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Rcpp_mgSEM.R
\name{AIC,Rcpp_mgSEM-method}
\alias{AIC,Rcpp_mgSEM-method}
\title{AIC}
\usage{
\S4method{AIC}{Rcpp_mgSEM}(object, ..., k = 2)
}
\arguments{
\item{object}{object of class Rcpp_mgSEM}

\item{...}{not used}

\item{k}{multiplier for number of parameters}
}
\value{
AIC values
}
\description{
AIC
}
