% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Rcpp_SEMCpp.R
\name{AIC,Rcpp_SEMCpp-method}
\alias{AIC,Rcpp_SEMCpp-method}
\title{AIC}
\usage{
\S4method{AIC}{Rcpp_SEMCpp}(object)
}
\arguments{
\item{object}{object of class Rcpp_SEMCpp}
}
\value{
AIC values
}
\description{
AIC
}
