% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-regularizedSEM.R
\name{AIC,regularizedSEM-method}
\alias{AIC,regularizedSEM-method}
\title{AIC}
\usage{
\S4method{AIC}{regularizedSEM}(object)
}
\arguments{
\item{object}{object of class regularizedSEM}
}
\value{
AIC values
}
\description{
returns the AIC
}
