% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-regularizedSEMWithCustomPenalty.R
\name{AIC,regularizedSEMWithCustomPenalty-method}
\alias{AIC,regularizedSEMWithCustomPenalty-method}
\title{AIC}
\usage{
\S4method{AIC}{regularizedSEMWithCustomPenalty}(object)
}
\arguments{
\item{object}{object of class regularizedSEMWithCustomPenalty}
}
\description{
returns the AIC
}
