% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularizeSEM.R
\name{alternativeStartingValues}
\alias{alternativeStartingValues}
\title{alternativeStartingValues}
\usage{
alternativeStartingValues(SEM)
}
\arguments{
\item{SEM}{model of class Rcpp_SEMCpp.}
}
\description{
generate alternative starting values for a SEM in raw format.
}
