% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMNET.R
\name{BFGS}
\alias{BFGS}
\title{BFGS}
\usage{
BFGS(
  oldParameters,
  oldGradients,
  oldHessian,
  newParameters,
  newGradients,
  cautious = TRUE,
  hessianEps = 0.001
)
}
\arguments{
\item{oldParameters}{parameters of previous iteration}

\item{oldGradients}{gradients of previous iteration}

\item{oldHessian}{Hessian of previous iteration}

\item{newParameters}{parameters of current iteration}

\item{newGradients}{gradients of current iteration}

\item{cautious}{boolean: should the update be skipped if it would result in a non positive definite Hessian?}

\item{hessianEps}{controls when the update of the Hessian approximation is skipped}
}
\description{
computes the BFGS Hessian approximation
}
