% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bfgsEnet}
\alias{bfgsEnet}
\title{smoothly approximated elastic net}
\value{
a list with fit results
}
\description{
Object for smoothly approximated elastic net optimization with
bfgs optimizer
}
\section{Fields}{

\describe{
\item{\code{new}}{creates a new object. Requires (1) a vector with weights for each
parameter and (2) a list with control elements}

\item{\code{setHessian}}{changes the Hessian of the model. Expects a matrix}

\item{\code{optimize}}{optimize the model. Expects a vector with starting values,
a SEM of type SEM_Cpp, a lambda and an alpha value.}
}}

