% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-approximateInfluence.R
\name{BIC,approximateInfluence-method}
\alias{BIC,approximateInfluence-method}
\title{BIC}
\usage{
\S4method{BIC}{approximateInfluence}(object)
}
\arguments{
\item{object}{object of class approximateInfluence}
}
\description{
returns the BIC
}
