% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-gpRegularized.R
\name{BIC,gpRegularized-method}
\alias{BIC,gpRegularized-method}
\title{BIC}
\usage{
\S4method{BIC}{gpRegularized}(object)
}
\arguments{
\item{object}{object of class gpRegularized}
}
\value{
data frame with fit values, appended with BIC
}
\description{
returns the BIC
}
