% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Rcpp_mgSEM.R
\name{BIC,Rcpp_mgSEM-method}
\alias{BIC,Rcpp_mgSEM-method}
\title{BIC}
\usage{
\S4method{BIC}{Rcpp_mgSEM}(object, ...)
}
\arguments{
\item{object}{object of class Rcpp_mgSEM}

\item{...}{not used}
}
\value{
BIC values
}
\description{
BIC
}
