% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Rcpp_SEMCpp.R
\name{BIC,Rcpp_SEMCpp-method}
\alias{BIC,Rcpp_SEMCpp-method}
\title{BIC}
\usage{
\S4method{BIC}{Rcpp_SEMCpp}(object)
}
\arguments{
\item{object}{object of class Rcpp_SEMCpp}
}
\value{
BIC values
}
\description{
BIC
}
