% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-regularizedSEM.R
\name{BIC,regularizedSEM-method}
\alias{BIC,regularizedSEM-method}
\title{BIC}
\usage{
\S4method{BIC}{regularizedSEM}(object)
}
\arguments{
\item{object}{object of class regularizedSEM}
}
\value{
BIC values
}
\description{
returns the BIC
}
