% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-regularizedSEMMixedPenalty.R
\name{BIC,regularizedSEMMixedPenalty-method}
\alias{BIC,regularizedSEMMixedPenalty-method}
\title{BIC}
\usage{
\S4method{BIC}{regularizedSEMMixedPenalty}(object)
}
\arguments{
\item{object}{object of class regularizedSEMMixedPenalty}
}
\value{
BIC values
}
\description{
returns the BIC
}
