% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-regularizedSEMWithCustomPenalty.R
\name{BIC,regularizedSEMWithCustomPenalty-method}
\alias{BIC,regularizedSEMWithCustomPenalty-method}
\title{BIC}
\usage{
\S4method{BIC}{regularizedSEMWithCustomPenalty}(object)
}
\arguments{
\item{object}{object of class regularizedSEMWithCustomPenalty}
}
\description{
returns the BIC
}
