% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{callFitFunction}
\alias{callFitFunction}
\title{callFitFunction}
\usage{
callFitFunction(fitFunctionSEXP, parameters, userSuppliedElements)
}
\arguments{
\item{fitFunctionSEXP}{pointer to fit function}

\item{parameters}{vector with parameter values}

\item{userSuppliedElements}{list with additional elements}
}
\value{
fit value (double)
}
\description{
wrapper to call user defined fit function
}
