% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cappedL1.R
\name{cappedElasticNet}
\alias{cappedElasticNet}
\title{cappedElasticNet}
\usage{
cappedElasticNet(
  lavaanModel,
  weights,
  lambdas,
  alphas,
  thetas,
  modifyModel = lessSEM::modifyModel(),
  control = controlIsta()
)
}
\arguments{
\item{lavaanModel}{model of class lavaan}

\item{weights}{labeled vector with weights for each of the parameters in the
model. If you are unsure what these parameters are called, use
getLavaanParameters(model) with your lavaan model object}

\item{lambdas}{numeric vector: values for the tuning parameter lambda}

\item{alphas}{numeric vector with values of the tuning parameter alpha. Must be
in \link{0,1}. 0 = ridge, 1 = lasso.}

\item{thetas}{parameters whose absolute value is above this threshold will be penalized with
a constant (theta)}

\item{modifyModel}{used to modify the lavaanModel. See ?modifyModel.}

\item{control}{used to control the optimizer. This element is generated with
the controlIsta().}
}
\description{
This function allows for regularization of models built in lavaan with a combination
of ridge and capped l1 penalty. THIS IS EXPERIMENTAL AND SHOULD NOT BE USED!
This capped L1 penalty is similar to the lasso but has an additional
tuning parameter (theta; see ?cappedL1). All parameter estimates above this penalty will be
penalized with constant theta.
The returned object is an S4 class; its elements can be accessed
with the "@" operator (see examples).
}
\details{
For more details, see:
\enumerate{
\item Gong, P., Zhang, C., Lu, Z., Huang, J., & Ye, J. (2013).
A general iterative shrinkage and thresholding algorithm for non-convex
regularized optimization problems. Proceedings of the 30th International
Conference on Machine Learning, 28(2)(2), 37–45.
}
}
\examples{
# see ?cappedL1, the use of cappedElasticNet is currently not recommended
}
