% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularizeSEM.R
\name{checkRegularizedParameters}
\alias{checkRegularizedParameters}
\title{checkRegularizedParameters}
\usage{
checkRegularizedParameters(
  parameters,
  regularizedParameterLabels,
  parameterTable,
  raw
)
}
\arguments{
\item{parameters}{labeled vector with parameter values}

\item{regularizedParameterLabels}{labels of regularized parameters}

\item{parameterTable}{table with all parameters}

\item{raw}{controls if the internal transformations of linr is used.}
}
\description{
internal function to check if the regularized parameters can be found in the model
}
