% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-aCV4RegularizedSEM.R
\name{coef,aCV4RegularizedSEM-method}
\alias{coef,aCV4RegularizedSEM-method}
\title{coef}
\usage{
\S4method{coef}{aCV4RegularizedSEM}(object, lambda = NULL, alpha = NULL)
}
\arguments{
\item{object}{object of class regularizedSEM}

\item{lambda}{numeric: value of lambda for which parameters should be returned. Must be present in object}

\item{alpha}{numeric: value of alpha for which parameters should be returned. Must be present in object}
}
\description{
Returns the parameter estimates of an aCV4RegularizedSEM
}
