% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-aCV4regularizedSEMWithCustomPenalty.R
\name{coef,aCV4regularizedSEMWithCustomPenalty-method}
\alias{coef,aCV4regularizedSEMWithCustomPenalty-method}
\title{coef}
\usage{
\S4method{coef}{aCV4regularizedSEMWithCustomPenalty}(object)
}
\arguments{
\item{object}{object of class regularizedSEM}
}
\description{
Returns the parameter estimates of an aCV4regularizedSEMWithCustomPenalty
}
