% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-approximateInfluence.R
\name{coef,approximateInfluenceFitIndices-method}
\alias{coef,approximateInfluenceFitIndices-method}
\title{coef}
\usage{
\S4method{coef}{approximateInfluenceFitIndices}(object, lambda = NULL, removedSubset = NULL)
}
\arguments{
\item{object}{object of class approximateInfluenceFitIndices}

\item{lambda}{for which lambda should the parameters be returned? If set to NULL, lambda is selected based on the information criterion}

\item{removedSubset}{for which removed subset should the parameters be returned? If set to NULL, parameters are returned for all removed subsets.}
}
\description{
Returns the parameter estimates for selected subsets. By default, the best parameter estimates for each subset are returned.
}
