% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CV4RegularizedSEM.R
\name{coef,CV4RegularizedSEM-method}
\alias{coef,CV4RegularizedSEM-method}
\title{coef}
\usage{
\S4method{coef}{CV4RegularizedSEM}(object, rule = "min")
}
\arguments{
\item{object}{object of class regularizedSEM}

\item{rule}{one of "min" or "1sd". With "min", the parameters of the model with the lowest CV-fit value will be returned.
With "1sd", the 1 standard deviation rule is used and the parameters with the larges lambda and a fit within 1 sd of the minimum will be returned.}

\item{lambda}{numeric: value of lambda for which parameters should be returned. Must be present in object}

\item{alpha}{numeric: value of alpha for which parameters should be returned. Must be present in object}
}
\description{
Returns the parameter estimates of an CV4RegularizedSEM
}
