% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cvRegularizedSEM.R
\name{coef,cvRegularizedSEM-method}
\alias{coef,cvRegularizedSEM-method}
\title{coef}
\usage{
\S4method{coef}{cvRegularizedSEM}(object, ...)
}
\arguments{
\item{object}{object of class cvRegularizedSEM}

\item{...}{not used}
}
\value{
the parameter estimates of an cvRegularizedSEM
}
\description{
Returns the parameter estimates of an cvRegularizedSEM
}
