% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Rcpp_mgSEM.R
\name{coef,Rcpp_mgSEM-method}
\alias{coef,Rcpp_mgSEM-method}
\title{coef}
\usage{
\S4method{coef}{Rcpp_mgSEM}(object, ...)
}
\arguments{
\item{object}{object of class Rcpp_mgSEM}
}
\value{
all coefficients of the model in transformed form
}
\description{
coef
}
