% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-regularizedSEM.R
\name{coef,regularizedSEM-method}
\alias{coef,regularizedSEM-method}
\title{coef}
\usage{
\S4method{coef}{regularizedSEM}(object, criterion = NULL, lambda = NULL, alpha = NULL)
}
\arguments{
\item{object}{object of class regularizedSEM}

\item{criterion}{can be one of: "AIC", "BIC". If set to NULL, all parameters will be returned}

\item{lambda}{numeric: value of lambda for which parameters should be returned. Must be present in object}

\item{alpha}{numeric: value of alpha for which parameters should be returned. Must be present in object}
}
\description{
Returns the parameter estimates of a regularizedSEM
}
