% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-regularizedSEMMixedPenalty.R
\name{coef,regularizedSEMMixedPenalty-method}
\alias{coef,regularizedSEMMixedPenalty-method}
\title{coef}
\usage{
\S4method{coef}{regularizedSEMMixedPenalty}(object, ...)
}
\arguments{
\item{object}{object of class regularizedSEMMixedPenalty}

\item{...}{criterion can be one of: "AIC", "BIC". If set to NULL, all parameters will be returned}
}
\value{
parameters of the model as data.frame
}
\description{
Returns the parameter estimates of a regularizedSEMMixedPenalty
}
