% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-regularizedSEMWithCustomPenalty.R
\name{coef,regularizedSEMWithCustomPenalty-method}
\alias{coef,regularizedSEMWithCustomPenalty-method}
\title{coef}
\usage{
\S4method{coef}{regularizedSEMWithCustomPenalty}(object, ...)
}
\arguments{
\item{object}{object of class regularizedSEMWithCustomPenalty}

\item{...}{not used}
}
\value{
data.frame with all parameter estimates
}
\description{
Returns the parameter estimates of a regularizedSEMWithCustomPenalty
}
\keyword{internal}
