% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeGroupM2LL}
\alias{computeGroupM2LL}
\title{computeGroupM2LL
Computes the -2log likelihood in a RAM model for a  group of people with identical 
missing structure. The idea is based on OpenMx (Boker, S. M., Neale, M. C., Maes, H. H., 
Wilde, M. J., Spiegel, M., Brick, T. R., Estabrook, R., Bates, T. C., & Mehta, P. (2022). 
OpenMx: Extended structural equation modelling (2.20.0). https://cran.r-project.org/web/packages/OpenMx/OpenMx.pdf)}
\usage{
computeGroupM2LL(
  sampleSize,
  nObservedVariables,
  observedMeans,
  observedCov,
  impliedMeans,
  impliedCovariance
)
}
\arguments{
\item{sampleSize}{number of persons}

\item{nObservedVariables}{number of non-missing variables}

\item{observedMeans}{vector with means of the observed variables}

\item{observedCov}{matrix with (co-)variances of the observed variables}

\item{impliedMeans}{implied means of the SEM}

\item{impliedCovariance}{implied covariance of the SEM}
}
\value{
-2 log-Lilekelihood
}
\description{
computeGroupM2LL
Computes the -2log likelihood in a RAM model for a  group of people with identical 
missing structure. The idea is based on OpenMx (Boker, S. M., Neale, M. C., Maes, H. H., 
Wilde, M. J., Spiegel, M., Brick, T. R., Estabrook, R., Bates, T. C., & Mehta, P. (2022). 
OpenMx: Extended structural equation modelling (2.20.0). https://cran.r-project.org/web/packages/OpenMx/OpenMx.pdf)
}
