% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeImpliedCovariance}
\alias{computeImpliedCovariance}
\title{computeImpliedCovariance}
\usage{
computeImpliedCovariance(Fmatrix, impliedCovarianceFull)
}
\arguments{
\item{Fmatrix}{filter matrix}

\item{impliedCovarianceFull}{implied covariance matrix including latent variables}
}
\value{
matrix with implied covariances
}
\description{
computes the implied covariance matrix of a SEM using RAM notation
}
