% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeImpliedCovarianceFull}
\alias{computeImpliedCovarianceFull}
\title{computeImpliedCovarianceFull}
\usage{
computeImpliedCovarianceFull(Amatrix, Smatrix, IminusAInverse)
}
\arguments{
\item{Amatrix}{matrix with directed effects}

\item{Smatrix}{matrix with undirected effects}

\item{IminusAInverse}{(I-Amatrix)^(-1)}
}
\value{
matrix with implied covariances
}
\description{
computes the implied covariance matrix including latent vaiables of a SEM using RAM notation
}
