% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeImpliedMeans}
\alias{computeImpliedMeans}
\title{computeImpliedMeans}
\usage{
computeImpliedMeans(Fmatrix, impliedMeansFull)
}
\arguments{
\item{Fmatrix}{filter matrix}

\item{impliedMeansFull}{implied means vector including latent variables}
}
\value{
matrix with implied means
}
\description{
computes the implied means vector of a SEM using RAM notation
}
