% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeImpliedMeansFull}
\alias{computeImpliedMeansFull}
\title{computeImpliedMeansFull}
\usage{
computeImpliedMeansFull(Amatrix, Mvector, IminusAInverse)
}
\arguments{
\item{Amatrix}{matrix with directed effects}

\item{Mvector}{vector with means}

\item{IminusAInverse}{(I-Amatrix)^(-1)}
}
\value{
matrix with implied means
}
\description{
computes the implied means vector of a SEM including the latent variables using RAM notation
}
