% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeIndividualM2LL}
\alias{computeIndividualM2LL}
\title{computeIndividualM2LL}
\usage{
computeIndividualM2LL(
  nObservedVariables,
  rawData,
  impliedMeans,
  impliedCovariance
)
}
\arguments{
\item{nObservedVariables}{number of non-missing variables}

\item{rawData}{observed data for this person}

\item{impliedMeans}{implied means of the SEM}

\item{impliedCovariance}{implied covariance of the SEM}
}
\value{
-2 log-Lilekelihood
}
\description{
Computes the -2log likelihood in a SEM for a single person
}
