% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEMData.R
\name{constructDataset}
\alias{constructDataset}
\title{constructDataset}
\usage{
constructDataset(rawData)
}
\arguments{
\item{rawData}{raw dataset with persons in rows and variables in columns}
}
\description{
Computes the number of unique missingness patterns
}
