% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlOptimizer.R
\name{controlQuasiNewtonBFGS}
\alias{controlQuasiNewtonBFGS}
\title{controlQuasiNewtonBFGS}
\usage{
controlQuasiNewtonBFGS(
  startingValues = "est",
  initialHessian = ifelse(all(startingValues == "est"), "compute", 1),
  stepSize = 1,
  c1 = 1e-04,
  c2 = 0.9,
  sig = 1e-05,
  gam = 0,
  maxIterOut = 1000,
  maxIterIn = 1000,
  maxIterLine = 500,
  epsOut = 1e-05,
  epsIn = 1e-05,
  useMultipleConvergenceCriteria = TRUE,
  regM2LLChangeEps = 1e-06,
  saveHessian = FALSE,
  verbose = 0
)
}
\arguments{
\item{startingValues}{option to provide initial starting values. Only used for the first lambda. Three options are supported. Setting to "est" will use the estimates
from the lavaan model object. Setting to "start" will use the starting values of the lavaan model. Finally, a labeled vector with parameter
values can be passed to the function which will then be used as starting values.}

\item{initialHessian}{option to provide an initial Hessian to the optimizer. Must have row and column names corresponding to the parameter labels. use getLavaanParameters(lavaanModel) to 
see those labels. If set to "compute", the initial hessian will be computed. If set to a single value, a diagonal matrix with the single value along the diagonal will be used.}

\item{stepSize}{Initial stepSize of the outer iteration (theta_{k+1} = theta_k + stepSize * Stepdirection)}

\item{sig}{only relevant when lineSearch = 'GLMNET'. Controls the sigma parameter in Yuan, G.-X., Ho, C.-H., & Lin, C.-J. (2012). An improved GLMNET for l1-regularized logistic regression. The Journal of Machine Learning Research, 13, 1999–2030. https://doi.org/10.1145/2020408.2020421.}

\item{gam}{Controls the gamma parameter in Yuan, G.-X., Ho, C.-H., & Lin, C.-J. (2012). An improved GLMNET for l1-regularized logistic regression. The Journal of Machine Learning Research, 13, 1999–2030. https://doi.org/10.1145/2020408.2020421. Defaults to 0.}

\item{maxIterOut}{Maximal number of outer iterations}

\item{maxIterIn}{Maximal number of inner iterations}

\item{maxIterLine}{Maximal number of iterations for the line search procedure}

\item{epsOut}{Stopping criterion for outer iterations}

\item{epsIn}{Stopping criterion for inner iterations}

\item{useMultipleConvergenceCriteria}{if set to TRUE, GLMNET will also check the change in fit and the change in parameters. If any convergence criterion is met, the optimization stops}

\item{regM2LLChangeEps}{if useMultipleConvergenceCriteria: change in fit which results in convergence}

\item{saveHessian}{should the Hessian be saved for later use? Note: This may take a lot of memory!}

\item{verbose}{0 prints no additional information, > 0 prints GLMNET iterations}
}
\description{
Control the quasiNewtonBFGS optimizer.
}
