% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSubsets.R
\name{createSubsets}
\alias{createSubsets}
\title{createSubsets}
\usage{
createSubsets(N, k)
}
\arguments{
\item{N}{number of samples in the data set}

\item{k}{number of subsets to create}
}
\value{
matrix with subsets
}
\description{
create subsets for cross-validation
}
\examples{
createSubsets(N=100, k = 5)
}
