% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aCV4model.R
\name{customACVRcpp_SEMCpp}
\alias{customACVRcpp_SEMCpp}
\title{customACVRcpp_SEMCpp}
\usage{
customACVRcpp_SEMCpp(
  SEM,
  subsets,
  raw,
  individualPenaltyFunction,
  individualPenaltyFunctionGradient,
  individualPenaltyFunctionHessian,
  currentTuningParameters,
  penaltyFunctionArguments
)
}
\arguments{
\item{SEM}{model of class Rcpp_SEMCpp. Models of this class
can be generated with the SEMFromLavaan-function.}

\item{subsets}{list with subsets created with createSubsets()}

\item{raw}{controls if the internal transformations of aCV4SEM should be used.}

\item{regularizedParameterLabels}{vector with labels of regularized parameters}

\item{lambda}{value of tuning parameter lambda}

\item{alpha}{value of tuning parameter alpha (for elastic net)}

\item{adaptiveLassoWeights}{vector with labeled adaptive lasso weights. Only required if penalty = "adaptiveLasso"}
}
\description{
internal function for approximate cross-validation based on the internal model representation of aCV4SEM. The custom part refers to the fact
that this function uses a custom penalty function.
}
