% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV4model.R
\name{CV4regularizedSEMWithCustomPenalty}
\alias{CV4regularizedSEMWithCustomPenalty}
\title{CV4regularizedSEMWithCustomPenalty}
\usage{
CV4regularizedSEMWithCustomPenalty(
  regularizedSEMWithCustomPenalty,
  k,
  returnSubsetParameters = FALSE
)
}
\arguments{
\item{k}{the number of cross-validation folds. Alternatively, 
a matrix with pre-defined subsets can be passed to the function. See ?linr::aCV4regularizedSEM for an example}

\item{returnSubsetParameters}{if set to TRUE, the parameter estimates of the individual cross-validation training sets will be returned}

\item{regularizedSEMWithCustomPenalty.}{model of class regularizedSEMWithCustomPenalty.}
}
\description{
Exact cross-validation for models of class regularizedSEMWithCustomPenalty. These models can be fit with regularizedSEMWithCustomPenalty.() (see ?regularizedSEMWithCustomPenalty.)
in this package.
}
\examples{
library(linr)

dataset <- simulateExampleData()

lavaanSyntax <- "
f =~ l1*y1 + l2*y2 + l3*y3 + l4*y4 + l5*y5 + 
     l6*y6 + l7*y7 + l8*y8 + l9*y9 + l10*y10 + 
     l11*y11 + l12*y12 + l13*y13 + l14*y14 + l15*y15
f ~~ 1*f
"

lavaanModel <- lavaan::sem(lavaanSyntax,
                           data = dataset,
                           meanstructure = TRUE,
                           std.lv = TRUE)

tuningParameters <- data.frame(lambda = seq(0,1,.1)) # we will test 11 lambdas here

penaltyFunctionArguments <- list(
  eps = 1e-10,
  regularizedParameterLabels = paste0("l", 6:15)
)

smoothLASSO <- function(
  # here are our three arguments:
  parameters,
  tuningParameters,
  penaltyFunctionArguments
){
  lambda <- tuningParameters$lambda 
  eps <- penaltyFunctionArguments$eps
  regularizedParameterLabels <- penaltyFunctionArguments$regularizedParameterLabels
  
  regularizedParameters <- parameters[regularizedParameterLabels]
  
  penaltyLasso <- lambda*sum(sqrt(regularizedParameters^2 + eps))
  
  return(penaltyLasso)
}

regsemApprox <- regularizeSEMWithCustomPenalty(lavaanModel = lavaanModel,
                                               individualPenaltyFunction = smoothLASSO,
                                               tuningParameters = tuningParameters,
                                               penaltyFunctionArguments = penaltyFunctionArguments)

## true cross-validation
CV <- CV4regularizedSEMWithCustomPenalty(regularizedSEMWithCustomPenalty = regsemApprox,
                                         k = 5)
plot(CV)
}
