% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv4elasticNet.R
\name{cv4ridge}
\alias{cv4ridge}
\title{cv4ridge}
\usage{
cv4ridge(regularizedSEM, k, returnSubsetParameters = FALSE)
}
\arguments{
\item{regularizedSEM}{model of class regularizedSEM}

\item{k}{the number of cross-validation folds. Alternatively, a matrix with pre-defined subsets can be passed to the function. 
See ?lessSEM::aCV4regularizedSEM for an example}

\item{returnSubsetParameters}{if set to TRUE, the parameter estimates of the individual cross-validation training sets will be returned}

\item{control}{parameters passed to the glmnet optimizer. Note that only arguments of the inner iteration are used. See ?controlGlmnet for more details}
}
\description{
Exact cross-validation for models of class regularizedSEM which have been fitted with ridge penalty
These models can be fit with ridge() (see ?ridge)
in this package.
}
\examples{
library(lessSEM)

}
