% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scad.R
\name{cv4scad}
\alias{cv4scad}
\title{cv4scad}
\usage{
cv4scad(
  regularizedSEM,
  k,
  dataSet = NULL,
  scaleData = FALSE,
  scalingFunction = function(dataSet, scalingArguments) scale(x = dataSet, center =
    scalingArguments$center, scale = scalingArguments$scale),
  scalingArguments = list(center = TRUE, scale = TRUE),
  returnSubsetParameters = FALSE
)
}
\arguments{
\item{regularizedSEM}{model of class regularizedSEM}

\item{k}{the number of cross-validation folds. Alternatively, a matrix with pre-defined subsets can be passed to the function. 
See ?lessSEM::aCV4regularizedSEM for an example}

\item{dataSet}{optional: Pass the full, unscaled data set to the function. 
This is important if the data has to be scaled prior to the analysis. If scaling
is performed on the full sample, this will result in dependencies between the 
subsets created by the cross-validation. To prevent this, pass the full data and use scaleData = TRUE,
the scalingFunction, and the scalingArguments}

\item{scaleData}{if set to TRUE, the subsets will be scaled using the scalingFunction}

\item{scalingFunction}{this function is used to scale the subsets. It MUST take two arguments:
first, the data set as matrix and second the scalingArguments. The latter can be anything you need
for the scaling}

\item{scalingArguments}{the second argument passed to scalingFunction. Can contain any number
of arguments needed for the scaling}

\item{returnSubsetParameters}{if set to TRUE, the parameter estimates of the individual cross-validation training sets will be returned}
}
\description{
Exact cross-validation for models of class regularizedSEM which have been fitted with scad penalty
These models can be fit with scad() (see ?scad)
in this package.
}
\examples{
# see ?scad

}
