% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cvRegularizedSEM.R
\docType{class}
\name{cvRegularizedSEM-class}
\alias{cvRegularizedSEM-class}
\title{Class for cross-validated regularized SEM}
\description{
Class for cross-validated regularized SEM
}
\section{Slots}{

\describe{
\item{\code{parameters}}{data.frame with parameter estimates for the best combination of the
tuning parameters}

\item{\code{transformations}}{transformed parameters}

\item{\code{cvfits}}{data.frame with all combinations of the
tuning parameters and the sum of the cross-validation fits}

\item{\code{parameterLabels}}{character vector with names of all parameters}

\item{\code{regularized}}{character vector with names of regularized parameters}

\item{\code{cvfitsDetails}}{data.frame with cross-validation fits for each subset}

\item{\code{subsets}}{matrix indicating which person is in which subset}

\item{\code{subsetParameters}}{optional: data.frame with parameter estimates for all
combinations of the tuning parameters in all subsets}

\item{\code{misc}}{list with additional return elements}
}}

