% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlOptimizer.R
\name{.adaptBreakingForWls}
\alias{.adaptBreakingForWls}
\title{.adaptBreakingForWls}
\usage{
.adaptBreakingForWls(lavaanModel, currentBreaking, selectedDefault)
}
\arguments{
\item{lavaanModel}{single model or vector of models}

\item{currentBreaking}{current breaking condition value}

\item{selectedDefault}{was default breaking condition selected?}
}
\value{
updated breaking
}
\description{
wls needs smaller breaking points than ml
}
