% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeSEM.R
\name{.addMeanStructure}
\alias{.addMeanStructure}
\title{.addMeanStructure}
\usage{
.addMeanStructure(parameterTable, manifestNames, MvectorElements)
}
\arguments{
\item{parameterTable}{table with parameters}

\item{manifestNames}{names of manifest variables}

\item{MvectorElements}{elements of the means vector}
}
\value{
parameterTable
}
\description{
adds a mean strucuture to the parameter table
}
\keyword{internal}
