% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupRegularizedSEM.R
\name{.checkLavaanModel}
\alias{.checkLavaanModel}
\title{.checkLavaanModel}
\usage{
.checkLavaanModel(lavaanModel)
}
\arguments{
\item{lavaanModel}{m0del of type lavaan}
}
\value{
nothing
}
\description{
checks model of type lavaan
}
\keyword{internal}
