% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedPenalty.R
\name{.checkPenalties}
\alias{.checkPenalties}
\title{.checkPenalties}
\usage{
.checkPenalties(mixedPenalty)
}
\arguments{
\item{mixedPenalty}{object of class mixedPenalty. This object can be created
with the mixedPenalty function. Penalties can be added with the addCappedL1,
addLasso, addLsp, addMcp, and addScad functions.}
}
\description{
Internal function to check a mixedPenalty object
}
