% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{.checkSyntax}
\alias{.checkSyntax}
\title{.checkSyntax}
\usage{
.checkSyntax(syntax)
}
\arguments{
\item{syntax}{syntax for parameter transformations}
}
\value{
nothing
}
\description{
check the syntax for parameter transformations
}
