% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupRegularizedSEM.R
\name{.createMultiGroupTransformations}
\alias{.createMultiGroupTransformations}
\title{.createMultiGroupTransformations}
\usage{
.createMultiGroupTransformations(transformations, parameterValues)
}
\arguments{
\item{transformations}{string with transformations}

\item{parameterValues}{values of parameters already in the model}
}
\value{
list with extended parameter vector and transformation function pointer
}
\description{
compiles the transformation function and adapts the parameter vector
}
\keyword{internal}
