% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeSEM.R
\name{.createParameterTable}
\alias{.createParameterTable}
\title{.createParameterTable}
\usage{
.createParameterTable(
  parameterValues,
  parameterLabels,
  modelParameters,
  parameterIDs
)
}
\arguments{
\item{parameterValues}{values of parameters}

\item{parameterLabels}{names of the parameters}

\item{modelParameters}{model parameters from lavaan}

\item{parameterIDs}{unique parameter ids from lavaan -> identify each parameter
with a unique number}
}
\value{
parameter table for lessSEM
}
\description{
create a parameter table using the elements extracted from lavaan
}
\keyword{internal}
