% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{.createRcppTransformationFunction}
\alias{.createRcppTransformationFunction}
\title{.createRcppTransformationFunction}
\usage{
.createRcppTransformationFunction(syntax, parameters)
}
\arguments{
\item{syntax}{syntax with user defined transformations}

\item{parameters}{labels of parameters used in these transformations}
}
\value{
string with functions for compilations with Rcpp
}
\description{
create an Rcpp function which uses the user-defined parameter transformation
}
\keyword{internal}
