% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupRegularizedSEM.R
\name{.createTransformations}
\alias{.createTransformations}
\title{.createTransformations}
\usage{
.createTransformations(transformations, parameterLabels, parameterTable)
}
\arguments{
\item{transformations}{string with transformations}

\item{parameterLabels}{labels of parameteres already in the model}

\item{parameterTable}{existing parameter table}
}
\value{
list with parameterTable and transformation function pointer
}
\description{
compiles the transformation function and adapts the parameterTable
}
\keyword{internal}
