% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaanHelperFunctions.R
\name{.cvregsem2LavaanParameters}
\alias{.cvregsem2LavaanParameters}
\title{.cvregsem2LavaanParameters}
\usage{
.cvregsem2LavaanParameters(cvregsemModel, lavaanModel)
}
\arguments{
\item{cvregsemModel}{model of class cvregsem}

\item{lavaanModel}{model of class lavaan}
}
\value{
regsem parameters with lavaan labels
}
\description{
helper function: regsem and lavaan use slightly different parameter labels. This function
can be used to translate the parameter labels of a cv_regsem object to lavaan labels
}
\keyword{internal}
