% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvRegularizeSmoothSEMInternal.R
\name{.cvRegularizeSmoothSEMInternal}
\alias{.cvRegularizeSmoothSEMInternal}
\title{.cvRegularizeSmoothSEMInternal}
\usage{
.cvRegularizeSmoothSEMInternal(
  lavaanModel,
  k,
  standardize,
  penalty,
  weights,
  returnSubsetParameters,
  tuningParameters,
  epsilon,
  modifyModel,
  method = "bfgs",
  control
)
}
\arguments{
\item{lavaanModel}{model of class lavaan}

\item{k}{the number of cross-validation folds. Alternatively, a matrix with pre-defined subsets can be passed to the function.
See ?lessSEM::cvSmoothLasso for an example}

\item{standardize}{should training and test sets be standardized?}

\item{penalty}{string: name of the penalty used in the model}

\item{weights}{labeled vector with weights for each of the parameters in the
model.}

\item{returnSubsetParameters}{if set to TRUE, the parameter estimates of the individual cross-validation training sets will be returned}

\item{tuningParameters}{data.frame with tuning parameter values}

\item{epsilon}{epsilon > 0; controls the smoothness of the approximation. Larger values = smoother}

\item{modifyModel}{used to modify the lavaanModel. See ?modifyModel.}

\item{method}{optimizer used. Currently only "bfgs" is supported.}

\item{control}{used to control the optimizer. This element is generated with
the controlBFGS function. See ?controlBFGS for more details.}
}
\value{
model of class cvRegularizedSEM
}
\description{
Combination of smoothly regularized structural equation model and cross-validation
}
\details{
Internal function: This function computes the regularized models
for all penalty functions which are implemented for bfgs.
Use the dedicated penalty functions (e.g., lessSEM::cvSmoothLasso) to penalize
the model.
}
\keyword{internal}
