% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeSEM.R
\name{.defineDerivatives}
\alias{.defineDerivatives}
\title{.defineDerivatives}
\usage{
.defineDerivatives(SEMList, parameterTable, modelMatrices)
}
\arguments{
\item{SEMList}{list representing SEM}

\item{parameterTable}{table with parameters}

\item{modelMatrices}{matrices of the RAM model}
}
\value{
SEMList
}
\description{
adds all elements required to compute the derivatives of the fitting function
with respect to the parameters to the SEMList
}
\keyword{internal}
