% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{.extractParametersFromSyntax}
\alias{.extractParametersFromSyntax}
\title{.extractParametersFromSyntax}
\usage{
.extractParametersFromSyntax(syntax)
}
\arguments{
\item{syntax}{syntax for parameter transformations}
}
\value{
vector with names of parameters used in the syntax
}
\description{
extract the names of the parameters in a syntax
}
