% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticDerivatives.R
\name{.getGradients}
\alias{.getGradients}
\title{.getGradients}
\usage{
.getGradients(SEM, raw)
}
\arguments{
\item{SEM}{model of class Rcpp_SEMCpp}

\item{raw}{controls if the internal transformations of lessSEM should be used.
lessSEM will use an exponential function for all variances to
avoid negative variances. When set to TRUE, the gradients will be given for the
internal parameter representation. Set to FALSE to get the usual
gradients}
}
\value{
vector with derivatives of the -2log-Likelihood with respect to each parameter
}
\description{
returns the gradients of a model of class Rcpp_SEMCpp. This is the internal model
representation. Models of this class
can be generated with the lessSEM:::.SEMFromLavaan-function.
}
\keyword{internal}
