% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMaxLambda.R
\name{.gpGetMaxLambda}
\alias{.gpGetMaxLambda}
\title{.gpGetMaxLambda}
\usage{
.gpGetMaxLambda(
  regularizedModel,
  par,
  fitFunction,
  gradientFunction,
  userSuppliedArguments,
  weights
)
}
\arguments{
\item{regularizedModel}{Model combining likelihood and lasso type penalty}

\item{par}{labeled vector with starting values}

\item{fitFunction}{R fit function}

\item{gradientFunction}{R gradient functions}

\item{userSuppliedArguments}{list with arguments for fitFunction and gradientFunction}

\item{weights}{weights given to each parameter in the penalty function}
}
\value{
first lambda value which sets all regularized parameters to zero (plus some tolerance)
}
\description{
generates a the first lambda which sets all regularized parameters to zero
}
\keyword{internal}
