% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeFit.R
\name{.gradientFunction}
\alias{.gradientFunction}
\title{.gradientFunction}
\usage{
.gradientFunction(par, SEM, raw)
}
\arguments{
\item{par}{labeled vector with parameter values}

\item{SEM}{model of class Rcpp_SEMCpp.}

\item{raw}{controls if the internal transformations of lessSEM is used.}
}
\value{
gradients of the model
}
\description{
internal function which returns the gradients of an object of class Rcpp_SEMCpp. This function can be used in optimizers
}
\keyword{internal}
