% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeFit.R
\name{.individualMinus2LogLikelihood}
\alias{.individualMinus2LogLikelihood}
\title{.individualMinus2LogLikelihood}
\usage{
.individualMinus2LogLikelihood(par, SEM, data, raw)
}
\arguments{
\item{par}{labeled vector with parameter values}

\item{SEM}{model of class Rcpp_SEMCpp.}

\item{data}{vector with data points for this single individual}

\item{raw}{controls if the internal transformations of lessSEM is used.}
}
\value{
-2 log Likelihood for each subject in the data set
}
\description{
internal function which returns the -2 log Likelihood for a single subject
}
