% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupRegularizedSEM.R
\name{.initializeMultiGroupSEMForRegularization}
\alias{.initializeMultiGroupSEMForRegularization}
\title{.initializeMultiGroupSEMForRegularization}
\usage{
.initializeMultiGroupSEMForRegularization(
  lavaanModels,
  startingValues,
  modifyModel
)
}
\arguments{
\item{lavaanModels}{vector with models of class lavaan}

\item{startingValues}{either set to est, start, or labeled vector with starting values}

\item{modifyModel}{user supplied model modifications}
}
\value{
model to be used by the regularization procedure
}
\description{
initializes the internal C++ SEM for regularization functions
}
\keyword{internal}
