% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupRegularizedSEM.R
\name{.initializeSEMForRegularization}
\alias{.initializeSEMForRegularization}
\title{.initializeSEMForRegularization}
\usage{
.initializeSEMForRegularization(lavaanModel, startingValues, modifyModel)
}
\arguments{
\item{lavaanModel}{model of class lavaan}

\item{startingValues}{either set to est, start, or labeled vector with starting values}

\item{modifyModel}{user supplied model modifications}
}
\value{
model to be used by the regularization procedure
}
\description{
initializes the internal C++ SEM for regularization functions
}
\keyword{internal}
