% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupRegularizedSEM.R
\name{.initializeWeights}
\alias{.initializeWeights}
\title{.initializeWeights}
\usage{
.initializeWeights(
  weights,
  penalty,
  method,
  createAdaptiveLassoWeights,
  control,
  lavaanModel,
  modifyModel,
  startingValues,
  rawParameters
)
}
\arguments{
\item{weights}{weight argument passed to function}

\item{penalty}{penalty used}

\item{createAdaptiveLassoWeights}{should adaptive lasso weights be created?}

\item{control}{list with control elements for optimizer}

\item{lavaanModel}{model of type lavaan}

\item{modifyModel}{list with model modifications}

\item{startingValues}{either set to est, start, or labeled vector with starting values}

\item{rawParameters}{raw parameters}
}
\value{
vector with weights
}
\description{
initialize the adaptive lasso weights
}
\keyword{internal}
