% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVignettes.R
\name{.knitVignettes}
\alias{.knitVignettes}
\title{.knitVignettes}
\usage{
.knitVignettes(dir = "vignettes")
}
\arguments{
\item{dir}{directory, where the vignettes are located.}
}
\value{
creates Rmd vignettes
}
\description{
Takes vignettes of format .lessmd and knits them to .Rmd files to be used
as vignettes. The Reason for this two-step approach is to reduce the runtime
on CRAN. The function is adapted from Stefan Kloppenborg at
https://www.kloppenborg.ca/2021/06/long-running-vignettes/
}
