% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularizationHelper.R
\name{.labelLavaanParameters}
\alias{.labelLavaanParameters}
\title{.labelLavaanParameters}
\usage{
.labelLavaanParameters(lavaanModel)
}
\arguments{
\item{lavaanModel}{fitted lavaan model}
}
\value{
parameterTable with labeled parameters
}
\description{
Adds labels to unlabeled parameters in the lavaan parameter table. Also
removes fixed parameters.
}
