% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaanHelperFunctions.R
\name{.lavaan2regsemLabels}
\alias{.lavaan2regsemLabels}
\title{.lavaan2regsemLabels}
\usage{
.lavaan2regsemLabels(lavaanModel)
}
\arguments{
\item{lavaanModel}{model of class lavaan}
}
\value{
a list with lavaan and regsem labels
}
\description{
helper function: regsem and lavaan use slightly different parameter labels. This function
can be used to get both sets of labels.
}
\keyword{internal}
