% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeFit.R
\name{.likelihoodRatioFit}
\alias{.likelihoodRatioFit}
\title{.likelihoodRatioFit}
\usage{
.likelihoodRatioFit(par, SEM, raw)
}
\arguments{
\item{par}{labeled vector with parameter values}

\item{SEM}{model of class Rcpp_SEMCpp.}

\item{raw}{controls if the internal transformations of lessSEM is used.}
}
\value{
likelihood ratio fit statistic
}
\description{
internal function which returns the likelihood ratio fit statistic
}
\keyword{internal}
