% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{.makeSingleLine}
\alias{.makeSingleLine}
\title{.makeSingleLine}
\usage{
.makeSingleLine(syntax, what)
}
\arguments{
\item{syntax}{reduced syntax}

\item{what}{which statement to look for (parameters or start)}
}
\value{
a syntax where multi-line statements are condensed to one line
}
\description{
checks if a parameter: or a start: statement spans multiple lines
and reduces it to one line.
}
\keyword{internal}
