% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.noDotDotDot}
\alias{.noDotDotDot}
\title{.noDotDotDot}
\usage{
.noDotDotDot(fn, fnName, ...)
}
\arguments{
\item{fn}{fit or gradient function. IMPORTANT: THE FIRST ARGUMENT TO
THE FUNCTION MUST BE THE PARAMETER VECTOR}

\item{fnName}{name of the function fn}

\item{...}{additional arguments}
}
\value{
list with (1) new function which wraps fn and (2) list with arguments passed to fn
}
\description{
remplaces the dot dot dot part of the fitting and gradient fuction
}
\keyword{internal}
