% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customRegularization.R
\name{.penaltyTypes}
\alias{.penaltyTypes}
\title{.penaltyTypes}
\usage{
.penaltyTypes(penalty)
}
\arguments{
\item{penalty}{either a number or the name of the penalty}
}
\description{
translates the penalty from a numeric value to the character or from
the character to the numeric value. The numeric value is used by the C++ backend.
}
\keyword{internal}
