% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{.reduceSyntax}
\alias{.reduceSyntax}
\title{.reduceSyntax}
\usage{
.reduceSyntax(syntax)
}
\arguments{
\item{syntax}{string with user defined transformations}
}
\value{
a cut and simplified version of the syntax
}
\description{
reduce user defined parameter transformation syntax to basic elements
}
\keyword{internal}
