% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penaltyFunctions.R
\name{.ridgeValue}
\alias{.ridgeValue}
\title{.ridgeValue}
\usage{
.ridgeValue(parameters, tuningParameters, penaltyFunctionArguments)
}
\arguments{
\item{parameters}{vector with labeled parameter values}

\item{tuningParameters}{list with field lambda (tuning parameter value)}

\item{penaltyFunctionArguments}{list with field regularizedParameterLabels (labels of regularized parameters)}
}
\value{
penalty function value
}
\description{
ridge penalty function
}
\keyword{internal}
