% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEMData.R
\name{.SEMdata}
\alias{.SEMdata}
\title{.SEMdata}
\usage{
.SEMdata(rawData)
}
\arguments{
\item{rawData}{matrix with raw data set}
}
\value{
list with internal representation of data
}
\description{
internal function. Creates internal data representation
}
\keyword{internal}
