% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEMData.R
\name{.SEMdataWLS}
\alias{.SEMdataWLS}
\title{.SEMdataWLS}
\usage{
.SEMdataWLS(rawData, lavaanModel)
}
\arguments{
\item{rawData}{matrix with raw data set}

\item{lavaanModel}{lavaan model}
}
\value{
list with internal representation of data
}
\description{
internal function. Creates internal data representation
}
\keyword{internal}
