% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeSEM.R
\name{.setAMatrix}
\alias{.setAMatrix}
\title{.setAMatrix}
\usage{
.setAMatrix(
  model,
  lavaanParameterTable,
  nLatent,
  nManifest,
  latentNames,
  manifestNames
)
}
\arguments{
\item{model}{model of class lavaan}

\item{lavaanParameterTable}{parameter table from lavaan}

\item{nLatent}{number of latent variables}

\item{nManifest}{number of manifest variables}

\item{latentNames}{names of latent variables}

\item{manifestNames}{names of manifest variables}
}
\description{
internal function. Populates the matrix with directed effects in RAM notation
}
\keyword{internal}
