% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeSEM.R
\name{.setFmatrix}
\alias{.setFmatrix}
\title{.setFmatrix}
\usage{
.setFmatrix(nManifest, manifestNames, nLatent, latentNames)
}
\arguments{
\item{nManifest}{number of manifest variables}

\item{manifestNames}{names of manifest variables}

\item{nLatent}{number of latent variables}

\item{latentNames}{names of latent variables}
}
\value{
matrix
}
\description{
returns the filter matrix of a RAM
}
\keyword{internal}
