% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeSEM.R
\name{.setMVector}
\alias{.setMVector}
\title{.setMVector}
\usage{
.setMVector(
  model,
  lavaanParameterTable,
  nLatent,
  nManifest,
  latentNames,
  manifestNames,
  rawData
)
}
\arguments{
\item{model}{model of class lavaan}

\item{lavaanParameterTable}{parameter table from lavaan}

\item{nLatent}{number of latent variables}

\item{nManifest}{number of manifest variables}

\item{latentNames}{names of latent variables}

\item{manifestNames}{names of manifest variables}

\item{rawData}{matrix with raw data}
}
\description{
internal function. Populates the vector with means in RAM notation
}
\keyword{internal}
