% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{.setParameters}
\alias{.setParameters}
\title{.setParameters}
\usage{
.setParameters(SEM, labels, values, raw)
}
\arguments{
\item{SEM}{model of class Rcpp_SEMCpp. Models of this class}

\item{labels}{vector with parameter labels}

\item{values}{vector with parameter values}

\item{raw}{are the parameters given in raw format or transformed?}
}
\value{
SEM with changed parameter values
}
\description{
change the parameters of the internal model representation.
}
\keyword{internal}
