% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlOptimizer.R
\name{.setupMulticore}
\alias{.setupMulticore}
\title{.setupMulticore}
\usage{
.setupMulticore(control)
}
\arguments{
\item{control}{object created with controlBFGS, controlIsta or controlGlmnet function}
}
\value{
nothing
}
\description{
setup for multi-core support
}
\keyword{internal}
