% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penaltyFunctions.R
\name{.smoothElasticNetGradient}
\alias{.smoothElasticNetGradient}
\title{.smoothElasticNetGradient}
\usage{
.smoothElasticNetGradient(
  parameters,
  tuningParameters,
  penaltyFunctionArguments
)
}
\arguments{
\item{parameters}{vector with labeled parameter values}

\item{tuningParameters}{list with fields lambda (tuning parameter value), alpha (0<alpha<1. Controls the weight of ridge and lasso terms. alpha = 1 is lasso, alpha = 0 ridge)}

\item{penaltyFunctionArguments}{list with fields regularizedParameterLabels (labels of regularized parameters), and eps (controls the smooth approximation of non-differential penalty functions (e.g., lasso, adaptive lasso, or elastic net). Smaller values result in closer approximation, but may also cause larger issues in optimization.)}
}
\value{
gradient values
}
\description{
smoothed version of non-differentiable elastic LASSO gradient
}
\keyword{internal}
