% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeFit.R
\name{.standardErrors}
\alias{.standardErrors}
\title{.standardErrors}
\usage{
.standardErrors(SEM, raw)
}
\arguments{
\item{SEM}{model of class Rcpp_SEMCpp.}

\item{raw}{controls if the internal transformations of lessSEM is used. If
set to TRUE, the standard errors will be returned for the internally used
parameter specification}
}
\value{
a vector with standard errors
}
\description{
compute the standard errors of a fitted SEM. IMPORTANT: Assumes that the
SEM has been fitted and the parameter estimates are at the ordinary maximum
likelihood estimates
}
\keyword{internal}
