% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedPenalty.R
\name{.useElasticNet}
\alias{.useElasticNet}
\title{.useElasticNet}
\usage{
.useElasticNet(mixedPenalty)
}
\arguments{
\item{mixedPenalty}{object of class mixedPenalty. This object can be created
with the mixedPenalty function. Penalties can be added with the addCappedL1,
addLasso, addLsp, addMcp, and addScad functions.}
}
\value{
TRUE if elastic net, FALSE otherwise
}
\description{
Internal function checking if elastic net is used
}
