% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticNet.R
\name{elasticNet}
\alias{elasticNet}
\title{elasticNet}
\usage{
elasticNet(
  lavaanModel,
  weights,
  lambdas = NULL,
  nLambdas = NULL,
  alphas,
  method = "ista",
  control = controlIsta()
)
}
\arguments{
\item{lavaanModel}{model of class lavaan}

\item{weights}{labeled vector with weights for each of the parameters in the 
model. If you are unsure what these parameters are called, use 
getLavaanParameters(<model>) with your lavaan model object}

\item{lambdas}{numeric vector: values for the tuning parameter lambda}

\item{nLambdas}{alternative to lambda: If alpha = 1, linr can automatically
compute the first lambda value which sets all regularized parameters to zero.
It will then generate nLambda values between 0 and the computed lambda.}

\item{alphas}{numeric vector with values of the tuning parameter alpha. Must be
in [0,1]. 0 = ridge, 1 = lasso.}

\item{method}{which optimizer should be used? Currently implemented are ista
and glmnet. With ista, the control argument can be used to switch to related procedures
(currently gist).}

\item{control}{used to control the optimizer. This element is generated with 
the controlIsta and controlGlmnet functions. See ?controlIsta and ?controlGlmnet
for more details.}
}
\description{
This function allows for regularization of models built in lavaan with the
elastic net penalty. See 
Zou, H., & Hastie, T. (2005). Regularization and variable selection via 
the elastic net. Journal of the Royal Statistical Society: 
Series B, 67(2), 301–320. https://doi.org/10.1111/j.1467-9868.2005.00503.x
for the details of this regularization technique.
}
