% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cvRegularizedSEM.R
\name{estimates,cvRegularizedSEM-method}
\alias{estimates,cvRegularizedSEM-method}
\title{estimates}
\usage{
\S4method{estimates}{cvRegularizedSEM}(object, criterion = NULL)
}
\arguments{
\item{object}{object of class cvRegularizedSEM}

\item{criterion}{not used}
}
\value{
returns a matrix with estimates
}
\description{
estimates
}
