% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-regularizedSEM.R
\name{estimates,regularizedSEM-method}
\alias{estimates,regularizedSEM-method}
\title{estimates}
\usage{
\S4method{estimates}{regularizedSEM}(object, criterion = NULL)
}
\arguments{
\item{object}{object of class regularizedSEM}

\item{criterion}{fit index (e.g., AIC) used to select the parameters}
}
\value{
returns a matrix with estimates
}
\description{
estimates
}
