% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-generics.R
\name{estimates}
\alias{estimates}
\title{S4 method to exract the estimates of an object}
\usage{
estimates(object, criterion = NULL, transformations = FALSE)
}
\arguments{
\item{object}{a model fitted with lessSEM}

\item{criterion}{fitIndice used to select the parameters}

\item{transformations}{boolean: Should transformations be returned?}
}
\value{
returns a matrix with estimates
}
\description{
S4 method to exract the estimates of an object
}
