% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-regularizedSEM.R
\name{fitIndices,regularizedSEM-method}
\alias{fitIndices,regularizedSEM-method}
\title{fitIndices}
\usage{
\S4method{fitIndices}{regularizedSEM}(object)
}
\arguments{
\item{object}{object of class regularizedSEM}
}
\value{
returns a data.frame with fit indices
}
\description{
fitIndices
}
