% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitIndices.R
\name{fitIndices,regularizedSEMMixedPenalty-method}
\alias{fitIndices,regularizedSEMMixedPenalty-method}
\title{fitIndices}
\usage{
\S4method{fitIndices}{regularizedSEMMixedPenalty}(object)
}
\arguments{
\item{object}{object of class regularizedSEMMixedPenalty}
}
\value{
returns a data.frame with fit indices
}
\description{
fitIndices
}
