% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-generics.R
\name{fitIndices}
\alias{fitIndices}
\title{S4 method to compute fit indices (e.g., AIC, BIC, ...)}
\usage{
fitIndices(object)
}
\arguments{
\item{object}{a model fitted with lessSEM}
}
\value{
returns a data.frame with fit indices
}
\description{
S4 method to compute fit indices (e.g., AIC, BIC, ...)
}
