% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penaltyFunctions.R
\name{genericGradientApproximiation}
\alias{genericGradientApproximiation}
\title{genericGradientApproximiation}
\usage{
genericGradientApproximiation(
  parameters,
  tuningParameters,
  penaltyFunctionArguments
)
}
\arguments{
\item{parameters}{vector with labeled parameter values}

\item{tuningParameters}{data.frame with tuning parameters}

\item{penaltyFunctionArguments}{list with additional arguments passed to the penalty function. NOTE: The penalty function itself must also be an object of this list! penalty function. Must accept three parameters: vector with parameter values, data.frame with tuning parameters, and list with penaltyFunctionArguments. see linr::smoothLASSO for an example}
}
\description{
This function can be used to approximate the gradients of a generic penalty function with numDeriv
}
