% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penaltyFunctions.R
\name{genericHessianApproximiation}
\alias{genericHessianApproximiation}
\title{genericHessianApproximiation}
\usage{
genericHessianApproximiation(
  parameters,
  tuningParameters,
  penaltyFunctionArguments
)
}
\arguments{
\item{parameters}{vector with labeled parameter values}

\item{tuningParameters}{data.frame with tuning parameters}

\item{penaltyFunctionArguments}{list with additional arguments passed to the penalty function. NOTE: The penalty function itself must also be an object of this list! penalty function. Must accept three parameters: vector with parameter values, data.frame with tuning parameters, and list with penaltyFunctionArguments. see aCV4SEM::smoothLASSO for an example}

\item{penaltyFunction}{}
}
\description{
This function can be used to approximate the Hessian of a generic penalty function with numDeriv
}
