% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticDerivatives.R
\name{getHessian}
\alias{getHessian}
\title{getHessian}
\usage{
getHessian(SEM, raw = FALSE, eps = 1e-07)
}
\arguments{
\item{SEM}{model of class Rcpp_SEMCpp}

\item{raw}{controls if the internal transformations of linr should be used. linr will use an exponential function for all variances to 
avoid negative variances. When set to TRUE, the gradients will be given for the internal parameter representation. Set to FALSE to get the usual 
gradients}

\item{eps}{eps controls the step size of the numerical approximation.}
}
\description{
returns the Hessian of a model of class Rcpp_SEMCpp. This is the internal model representation. Models of this class
can be generated with the SEMFromLavaan-function. The function is adapted from \link[lavaan]{lav_model_hessian}.
}
