% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaanHelperFunctions.R
\name{getLavaanParameters}
\alias{getLavaanParameters}
\title{getLavaanParameters}
\usage{
getLavaanParameters(lavaanModel, removeDuplicates = TRUE)
}
\arguments{
\item{lavaanModel}{model of class lavaan}

\item{removeDuplicates}{should duplicated parameters be removed?}
}
\value{
returns a labeled vector with parameters from lavaan
}
\description{
helper function: returns a labeled vector with parameters from lavaan
}
