% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularizeSEM.R
\name{getMaxLambda}
\alias{getMaxLambda}
\title{getMaxLambda}
\usage{
getMaxLambda(
  SEM,
  regularizedParameterLabels,
  adaptiveLassoWeights,
  initialHessian4Optimizer,
  control,
  N
)
}
\arguments{
\item{SEM}{model of class Rcpp_SEMCpp}

\item{regularizedParameterLabels}{labels of regularized parameters}

\item{adaptiveLassoWeights}{vector with weights for adaptive LASSO}

\item{initialHessian4Optimizer}{initial Hessian used by the optimizer}

\item{control}{additional arguments passed to GLMNET. See ?controlGLMNET}

\item{N}{sample size}
}
\description{
generates a the first lambda which sets all regularized parameters to zero
}
