% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMaxLambda.R
\name{getMaxLambda_C}
\alias{getMaxLambda_C}
\title{getMaxLambda_C}
\usage{
getMaxLambda_C(
  regularizedModel,
  SEM,
  rawParameters,
  weights,
  N,
  approx = FALSE
)
}
\arguments{
\item{regularizedModel}{Model combining likelihood and lasso type penalty}

\item{SEM}{model of class Rcpp_SEMCpp}

\item{rawParameters}{labeled vector with starting values}

\item{weights}{weights given to each parameter in the penalty function}

\item{N}{sample size}
}
\description{
generates a the first lambda which sets all regularized parameters to zero
}
