% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaanHelperFunctions.R
\name{getParameters.lavaan}
\alias{getParameters.lavaan}
\title{getParameters.lavaan}
\usage{
getParameters.lavaan(lavaanModel, removeDuplicates = TRUE)
}
\arguments{
\item{lavaanModel}{model of class lavaan}

\item{removeDuplicates}{should duplicated parameters be removed?}
}
\description{
helper function: returns a labeled vector with parameters from lavaan
}
