% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMNET.R
\name{getSubgradients}
\alias{getSubgradients}
\title{getSubgradients}
\usage{
getSubgradients(
  parameters,
  gradients,
  lambda,
  alpha,
  regularizedParameterLabels,
  adaptiveLassoWeights
)
}
\arguments{
\item{parameters}{vector with labeled parameter values}

\item{gradients}{gradients of the differentiable part}

\item{lambda}{lasso tuning parameter. Higher values = higher penalty}

\item{alpha}{0<alpha<1. Controls the weight of ridge and lasso terms. alpha = 1 is lasso, alpha = 0 ridge}

\item{regularizedParameterLabels}{labels of regularized parameters}

\item{adaptiveLassoWeights}{vector with weights for adaptive LASSO. Set to vector of ones if not using adaptive LASSO}
}
\description{
returns the sub-gradients at the current position
}
