% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-regularizedSEMMixedPenalty.R
\name{getTuningParameterConfiguration}
\alias{getTuningParameterConfiguration}
\title{getTuningParameterConfiguration}
\usage{
getTuningParameterConfiguration(
  regularizedSEMMixedPenalty,
  tuningParameterConfiguration
)
}
\arguments{
\item{regularizedSEMMixedPenalty}{object of type regularizedSEMMixedPenalty (see ?mixedPenalty)}

\item{tuningParameterConfiguration}{integer indicating which tuningParameterConfiguration
should be extracted (e.g., 1). See the entry in the row tuningParameterConfiguration of
regularizedSEMMixedPenalty@fits and regularizedSEMMixedPenalty@parameters.}
}
\value{
data frame with penalty and tuning parameter settings
}
\description{
Returns the lambda, theta, and alpha values for the tuning parameters
of a regularized SEM with mixed penalty.
}
