% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIC.R
\name{GIC}
\alias{GIC}
\title{GIC}
\usage{
GIC(regularizedSEM, scaler = 2)
}
\arguments{
\item{regularizedSEM}{model of class regularizedSEM}

\item{scaler}{numeric value to scale the degrees of freedom}
}
\value{
vector with GIC values.
}
\description{
THIS FUNCTION IS UNDER DEVELOPMENT AND SHOULD \emph{NOT} BE USED.
computes the generalized information criterion as
\eqn{-2\log \mathcal L + \text{scaler}*\text{df}}
where scaler is a numeric value with which the
degrees of freedom (df) are multiplied.
}
\details{
See Zhang et al. (2010) for more details.
\itemize{
\item Zhang, Y., Li, R., & Tsai, C.-L. (2010). Regularization Parameter
Selections via Generalized Information Criterion.
Journal of the American Statistical Association, 105(489),
312–323. https://doi.org/10.1198/jasa.2009.tm08013
}
}
