% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMNET.R
\name{GLMNET}
\alias{GLMNET}
\title{GLMNET}
\usage{
GLMNET(
  SEM,
  regularizedParameterLabels,
  lambda,
  alpha,
  adaptiveLassoWeights,
  initialHessian = NULL,
  stepSize = 1,
  c1 = 1e-04,
  c2 = 0.9,
  sig = 10^(-5),
  gam = 0,
  maxIterOut = 100,
  maxIterIn = 1000,
  maxIterLine = 500,
  epsOut = 1e-06,
  epsIn = 1e-06,
  verbose = 0
)
}
\arguments{
\item{SEM}{model of class Rcpp_SEMCpp.}

\item{regularizedParameterLabels}{labels of regularized parameters}

\item{lambda}{lasso tuning parameter. Higher values = higher penalty}

\item{alpha}{0<alpha<1. Controls the weight of ridge and lasso terms. alpha = 1 is lasso, alpha = 0 ridge, in between is elastic net}

\item{adaptiveLassoWeights}{vector with weights for adaptive LASSO. Set to NULL if not using adaptive LASSO}

\item{stepSize}{Initial stepSize of the outer iteration (theta_{k+1} = theta_k + stepSize * Stepdirection)}

\item{c1}{c1 constant for lineSearch. This constant controls the Armijo condition in lineSearch if lineSearch = "Wolfe"}

\item{c2}{c2 constant for lineSearch. This constant controls the Curvature condition in lineSearch if lineSearch = "Wolfe"}

\item{sig}{only relevant when lineSearch = 'GLMNET'. Controls the sigma parameter in Yuan, G.-X., Ho, C.-H., & Lin, C.-J. (2012). An improved GLMNET for l1-regularized logistic regression. The Journal of Machine Learning Research, 13, 1999–2030. https://doi.org/10.1145/2020408.2020421.}

\item{gam}{Controls the gamma parameter in Yuan, G.-X., Ho, C.-H., & Lin, C.-J. (2012). An improved GLMNET for l1-regularized logistic regression. The Journal of Machine Learning Research, 13, 1999–2030. https://doi.org/10.1145/2020408.2020421. Defaults to 0.}

\item{maxIterOut}{Maximal number of outer iterations}

\item{maxIterIn}{Maximal number of inner iterations}

\item{maxIterLine}{Maximal number of iterations for the line search procedure}

\item{epsOut}{Stopping criterion for outer iterations}

\item{epsIn}{Stopping criterion for inner iterations}
}
\description{
Performs GLMNET (see Friedman, 2010 & Yuan, 2011) with bfgs approximated Hessian
}
