% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximateCrossValidation.R
\name{GLMNETACVRcpp_SEMCpp}
\alias{GLMNETACVRcpp_SEMCpp}
\title{GLMNETACVRcpp_SEMCpp}
\usage{
GLMNETACVRcpp_SEMCpp(
  SEM,
  subsets,
  raw = FALSE,
  regularizedParameterLabels,
  lambda,
  alpha = NULL,
  adaptiveLassoWeights = NULL,
  hessianOfDifferentiablePart = NULL,
  maxIter = 100,
  epsBreak = 1e-10
)
}
\arguments{
\item{SEM}{model of class Rcpp_SEMCpp. Models of this class
can be generated with the SEMFromLavaan-function.}

\item{subsets}{list with subsets created with createSubsets()}

\item{raw}{controls if the internal transformations of aCV4SEM should be used.}

\item{regularizedParameterLabels}{vector with labels of regularized parameters}

\item{lambda}{value of tuning parameter lambda}

\item{alpha}{value of tuning parameter alpha (for elastic net)}

\item{adaptiveLassoWeights}{vector with labeled adaptive lasso weights. Only required if penalty = "adaptiveLasso"}

\item{maxIter}{maximal number of iterations used in GLMENT}

\item{epsBreak}{breaking condition of GLMNET}
}
\description{
internal function for approximate cross-validation based on the internal model representation of aCV4SEM. The GLMNET part refers to the fact
that this function uses the GLMNET optimizer for non-differentiable penalty functions.
}
