% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aI4model.R
\name{GLMNETApproximateInfluenceRcpp_SEMCpp}
\alias{GLMNETApproximateInfluenceRcpp_SEMCpp}
\title{GLMNETApproximateInfluenceRcpp_SEMCpp}
\usage{
GLMNETApproximateInfluenceRcpp_SEMCpp(
  SEM,
  subsets,
  raw = FALSE,
  regularizedParameterLabels,
  lambda,
  alpha = NULL,
  adaptiveLassoWeights = NULL,
  hessianOfDifferentiablePart = NULL,
  control
)
}
\arguments{
\item{SEM}{model of class Rcpp_SEMCpp. Models of this class
can be generated with the SEMFromLavaan-function.}

\item{subsets}{list with subsets created with createSubsets()}

\item{raw}{controls if the internal transformations of lessSEM should be used.}

\item{regularizedParameterLabels}{vector with labels of regularized parameters}

\item{lambda}{value of tuning parameter lambda}

\item{alpha}{value of tuning parameter alpha (for elastic net)}

\item{adaptiveLassoWeights}{vector with labeled adaptive lasso weights. Only required if penalty = "adaptiveLasso"}

\item{control}{parameters passed to the GLMNET optimizer. Note that only arguments of the inner iteration are used. See ?controlGLMNET for more details}

\item{maxIter}{maximal number of iterations used in GLMENT}

\item{epsBreak}{breaking condition of GLMNET}
}
\description{
internal function for approximate influence based on the internal model representation of lessSEM. The GLMNET part refers to the fact
that this function uses the GLMNET optimizer for non-differentiable penalty functions.
}
