% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{glmnetEnetGeneralPurposeCpp}
\alias{glmnetEnetGeneralPurposeCpp}
\title{elastic net optimization with glmnet optimizer}
\value{
a list with fit results
}
\description{
Object for elastic net optimization with
glmnet optimizer
}
\section{Fields}{

\describe{
\item{\code{new}}{creates a new object. Requires (1) a vector with weights for each
parameter and (2) a list with control elements}

\item{\code{setHessian}}{changes the Hessian of the model. Expects a matrix}

\item{\code{optimize}}{optimize the model. Expects a vector with starting values,
a SEXP function pointer to compute the fit, a SEXP function pointer to compute the gradients, a
list with elements the fit and gradient function require, a lambda and an alpha value.}
}}

