% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{glmnetLspMgSEM}
\alias{glmnetLspMgSEM}
\title{lsp optimization with glmnet optimizer}
\value{
a list with fit results
}
\description{
Object for lsp optimization with
glmnet optimizer
}
\section{Fields}{

\describe{
\item{\code{new}}{creates a new object. Requires (2) a list with control elements}

\item{\code{setHessian}}{changes the Hessian of the model. Expects a matrix}

\item{\code{optimize}}{optimize the model. Expects a vector with starting values,
a SEM of type SEM_Cpp, a theta and a lambda value.}
}}

