% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-gpRegularized.R
\docType{class}
\name{gpRegularized-class}
\alias{gpRegularized-class}
\title{Class for regularized model using general purpose optimization interface}
\description{
Class for regularized model using general purpose optimization interface
}
\section{Slots}{

\describe{
\item{\code{penalty}}{penalty used (e.g., "lasso")}

\item{\code{parameters}}{data.frame with all parameter estimates}

\item{\code{fits}}{data.frame with all fit results}

\item{\code{parameterLabels}}{character vector with names of all parameters}

\item{\code{weights}}{vector with weights given to each of the parameters in the penalty}

\item{\code{regularized}}{character vector with names of regularized parameters}

\item{\code{internalOptimization}}{list of elements used internally}

\item{\code{inputArguments}}{list with elements passed by the user to the general
purpose optimizer}
}}

