% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{istaCappedL1GeneralPurposeCpp}
\alias{istaCappedL1GeneralPurposeCpp}
\title{cappedL1 optimization with ista}
\value{
a list with fit results
}
\description{
Object for cappedL1 optimization with
ista optimizer
}
\section{Fields}{

\describe{
\item{\code{new}}{creates a new object. Requires (1) a vector with weights for each
parameter and (2) a list with control elements}

\item{\code{optimize}}{optimize the model. Expects a vector with starting values,
a SEXP function pointer to compute the fit, a SEXP function pointer to compute the gradients, a
list with elements the fit and gradient function require, a theta, lambda and an alpha value (alpha must be 1).}
}}

