% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ista.R
\name{istaLasso}
\alias{istaLasso}
\title{istaLasso}
\usage{
istaLasso(
  SEM,
  lambda,
  weights,
  i_k = 2,
  L0 = 1,
  eta = 2,
  maxIterOut = 1000,
  maxIterIn = 1000,
  breakOuter = 1e-07,
  raw = FALSE
)
}
\arguments{
\item{SEM}{object of class SEMCpp}

\item{lambda}{tuning parameter lambda}

\item{weights}{weights given to penalty for each parameter}

\item{i_k}{currently not used}

\item{L0}{controls the initial step size and change in step size (L0*(eta^k), where k is the iteration)}

\item{eta}{controls change of step size in ista (L0*(eta^k), where k is the iteration)}

\item{maxIterOut}{maximal number of outer iterations}

\item{maxIterIn}{maximal number of inner iterations}

\item{breakOuter}{change in fit below which convergence is assumend}

\item{raw}{should parameters be returned in raw format?}
}
\description{
Optimize LASSO regularized SEM with ista optimizer
}
