% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{istaMixedPenaltyGeneralPurpose}
\alias{istaMixedPenaltyGeneralPurpose}
\title{mixed penalty optimization with ista}
\value{
a list with fit results
}
\description{
Object for elastic net optimization with
ista optimizer
}
\section{Fields}{

\describe{
\item{\code{new}}{creates a new object.}

\item{\code{optimize}}{optimize the model.}
}}

