% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{istaMixedPenaltyGeneralPurposeCpp}
\alias{istaMixedPenaltyGeneralPurposeCpp}
\title{mixed penalty optimization with ista}
\value{
a list with fit results
}
\description{
Object for elastic net optimization with
ista optimizer
}
\section{Fields}{

\describe{
\item{\code{new}}{creates a new object. Requires (1) a vector with weights for each
parameter, (2) a vector indicating which penalty is used, and (3) a list with control elements}

\item{\code{optimize}}{optimize the model.}
}}

