% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticNet.R
\name{lasso}
\alias{lasso}
\title{lasso}
\usage{
lasso(
  lavaanModel,
  regularized,
  lambdas = NULL,
  nLambdas = NULL,
  method = "ista",
  control = controlIsta()
)
}
\arguments{
\item{lavaanModel}{model of class lavaan}

\item{regularized}{vector with names of parameters which are to be regularized.
If you are unsure what these parameters are called, use 
getLavaanParameters(<model>) with your lavaan model object}

\item{lambdas}{numeric vector: values for the tuning parameter lambda}

\item{nLambdas}{alternative to lambda: If alpha = 1, linr can automatically
compute the first lambda value which sets all regularized parameters to zero.
It will then generate nLambda values between 0 and the computed lambda.}

\item{method}{which optimizer should be used? Currently implemented are ista
and glmnet. With ista, the control argument can be used to switch to related procedures
(currently gist).}

\item{control}{used to control the optimizer. This element is generated with 
the controlIsta and controlGlmnet functions. See ?controlIsta and ?controlGlmnet
for more details.}
}
\description{
This function allows for regularization of models built in lavaan with the
lasso penalty. See, among others:
1) Tibshirani, R. (1996). Regression Shrinkage and Selection via the Lasso. 
Journal of the Royal Statistical Society. Series B (Methodological), 58(1), 267–288.
2) Jacobucci, R., Grimm, K. J., & McArdle, J. J. (2016). 
Regularized Structural Equation Modeling. Structural Equation Modeling: 
A Multidisciplinary Journal, 23(4), 555–566. https://doi.org/10.1080/10705511.2016.1154793
3) Huang, P.-H., Chen, H., & Weng, L.-J. (2017). A Penalized Likelihood 
Method for Structural Equation Modeling. Psychometrika, 82(2),
 329–354. https://doi.org/10.1007/s11336-017-9566-9
}
