% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaanHelperFunctions.R
\name{lavaan2lslxLabels}
\alias{lavaan2lslxLabels}
\title{lavaan2lslxLabels}
\usage{
lavaan2lslxLabels(lavaanModel)
}
\arguments{
\item{lavaanModel}{model of class lavaan}
}
\value{
list with lavaan labels and lslx labels
}
\description{
helper function: lslx and lavaan use slightly different parameter labels. This function
can be used to get both sets of labels.
}
