% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaanHelperFunctions.R
\name{lessSEM2Lavaan}
\alias{lessSEM2Lavaan}
\title{lessSEM2Lavaan}
\usage{
lessSEM2Lavaan(regularizedSEM, lambda, theta = NULL)
}
\arguments{
\item{regularizedSEM}{object created with lessSEM}

\item{lambda}{value for tuning parameter lambda}

\item{theta}{value for tuning parameter theta}
}
\value{
lavaan model
}
\description{
Creates a lavaan model object from lessSEM (only if possible).
}
