% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-coef.R
\docType{class}
\name{lessSEMCoef-class}
\alias{lessSEMCoef-class}
\title{Class for the coefficients estimated by lessSEM.}
\description{
Class for the coefficients estimated by lessSEM.
}
\section{Slots}{

\describe{
\item{\code{tuningParameters}}{tuning parameters}

\item{\code{estimates}}{parameter estimates}
}}

