% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logicalMatch}
\alias{logicalMatch}
\title{logicalMatch}
\usage{
logicalMatch(X, x)
}
\arguments{
\item{X}{matrix with booleans}

\item{x}{vector of booleans}
}
\value{
numerical vector with indices of matching rows
}
\description{
Returns the rows for which all elements of a boolean matrix X are equal
to the elements in boolean vector x
}
