% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Rcpp_mgSEM.R
\name{logLik,Rcpp_mgSEM-method}
\alias{logLik,Rcpp_mgSEM-method}
\title{logLik}
\usage{
\S4method{logLik}{Rcpp_mgSEM}(object)
}
\arguments{
\item{object}{object of class Rcpp_mgSEM}
}
\value{
log-likelihood of the model
}
\description{
logLik
}
