% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Rcpp_SEMCpp.R
\name{logLik,Rcpp_SEMCpp-method}
\alias{logLik,Rcpp_SEMCpp-method}
\title{logLik}
\usage{
\S4method{logLik}{Rcpp_SEMCpp}(object)
}
\arguments{
\item{object}{object of class Rcpp_SEMCpp}
}
\value{
log-likelihood of the model
}
\description{
logLik
}
