% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-logLikelihood.R
\docType{class}
\name{logLikelihood-class}
\alias{logLikelihood-class}
\title{Class for log-likelihood of regularized SEM. Note: we define a custom logLik -
Function because the generic one is using df = number of parameters which might be confusing.}
\description{
Class for log-likelihood of regularized SEM. Note: we define a custom logLik -
Function because the generic one is using df = number of parameters which might be confusing.
}
\section{Slots}{

\describe{
\item{\code{logLik}}{log-Likelihood}

\item{\code{nParameters}}{number of parameters in the model}

\item{\code{N}}{number of persons in the data set}
}}

