% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makePtrs}
\alias{makePtrs}
\title{makePtrs}
\usage{
makePtrs(fitFunName, gradFunName)
}
\arguments{
\item{fitFunName}{name of your C++ fit function (IMPORTANT: This must be the name
used in C++)}

\item{gradFunName}{name of your C++ gradient function (IMPORTANT: This must be the name
used in C++)}
}
\description{
This function helps you create the pointers necessary to use the Cpp interface
}
