% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mcpPenalty_C}
\alias{mcpPenalty_C}
\title{mcpPenalty_C}
\usage{
mcpPenalty_C(par, lambda_p, theta)
}
\arguments{
\item{par}{single parameter value}

\item{lambda_p}{lambda value for this parameter}

\item{theta}{theta value for this parameter}
}
\value{
penalty value
}
\description{
mcpPenalty_C
}
