% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-approximateInfluence.R
\name{mean,approximateInfluenceFitIndices-method}
\alias{mean,approximateInfluenceFitIndices-method}
\title{mean}
\usage{
\S4method{mean}{approximateInfluenceFitIndices}(x)
}
\arguments{
\item{x}{object of class approximateInfluenceFitIndices}
}
\description{
Returns the mean information criterion value over all subsets
}
