% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-approximateInfluence.R
\name{median,approximateInfluenceFitIndices-method}
\alias{median,approximateInfluenceFitIndices-method}
\title{median}
\usage{
\S4method{median}{approximateInfluenceFitIndices}(x, na.rm = NULL)
}
\arguments{
\item{x}{object of class approximateInfluenceFitIndices}
}
\description{
Returns the median information criterion value over all subsets
}
