% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mgSEM}
\alias{mgSEM}
\title{mgSEM class}
\description{
internal mgSEM representation
}
\section{Fields}{

\describe{
\item{\code{new}}{Creates a new mgSEM.}

\item{\code{addModel}}{add a model. Expects Rcpp::List}

\item{\code{addTransformation}}{adds transforamtions to a model}

\item{\code{implied}}{Computes implied means and covariance matrix}

\item{\code{fit}}{Fits the model. Returns objective value of the fitting function}

\item{\code{getParameters}}{Returns a data frame with model parameters.}

\item{\code{getParameterLabels}}{Returns a vector with unique parameter labels as used internally.}

\item{\code{getEstimator}}{Returns a vector with names of the estimators used in the submodels.}

\item{\code{getGradients}}{Returns a matrix with scores.}

\item{\code{getScores}}{Returns a matrix with scores. Not yet implemented}

\item{\code{getHessian}}{Returns the hessian of the model. Expects the labels of the
parameters and the values of the parameters as well as a boolean indicating if
these are raw. Finally, a double (eps) controls the precision of the approximation.}

\item{\code{computeTransformations}}{compute the transformations.}

\item{\code{setTransformationGradientStepSize}}{change the step size of the gradient computation for the transformations}
}}

