% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularizedSmoothSEMInternal.R
\name{newTau}
\alias{newTau}
\title{newTau}
\usage{
newTau(regularizedSEM, tau)
}
\arguments{
\item{regularizedSEM}{object fitted with approximate optimization}

\item{tau}{new tau value}
}
\value{
regularizedSEM, but with new regularizedSEM@fits$nonZeroParameters
}
\description{
assign new value to parameter tau used by approximate optimization. Any regularized
value below tau will be evaluated as zeroed which directly impacts the AIC, BIC, etc.
}
