% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularizeSEM.R
\name{optimizeCustomRegularizedSEM}
\alias{optimizeCustomRegularizedSEM}
\title{optimizeCustomRegularizedSEM}
\usage{
optimizeCustomRegularizedSEM(
  SEM,
  individualPenaltyFunction,
  individualPenaltyFunctionGradient,
  penaltyFunctionArguments,
  raw = TRUE,
  method = "BFGS",
  control = list(),
  hessian = FALSE
)
}
\arguments{
\item{SEM}{model of class Rcpp_SEMCpp.}

\item{individualPenaltyFunction}{penalty function which takes the current parameter values as first argument and the penaltyFunctionArguments as second argument and 
returns a single value - the value of the penalty function for a single person. If the true penalty function is non-differentiable (e.g., lasso) a smooth
approximation of this function should be provided.}

\item{individualPenaltyFunctionGradient}{gradients of the penalty function. Function should take the current parameter values as first argument and the penaltyFunctionArguments as second argument and 
return a vector of the same length as parameters. If the true penalty function is non-differentiable (e.g., lasso) a smooth
approximation of this function should be provided.}

\item{raw}{controls if the internal transformations of aCV4SEM is used.}

\item{method}{optimizer method. See ?optim}

\item{control}{control optimizer. See ?optim}

\item{hessian}{Should the optimizer return the hessian?}

\item{individualPenaltyFunctionHessian}{Hessian of the penalty function. Function should take the current parameter values as first argument and the penaltyFunctionArguments as second argument and 
return a matrix with (length as parameters)^2 number of elements. If the true penalty function is non-differentiable (e.g., lasso) a smooth
approximation of this function should be provided.}
}
\description{
optimize a SEM with custom penalty function.
}
