% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularizeSEM.R
\name{optimizeSEM}
\alias{optimizeSEM}
\title{optimizeSEM}
\usage{
optimizeSEM(
  SEM,
  raw = TRUE,
  method = "BFGS",
  control = list(),
  hessian = FALSE
)
}
\arguments{
\item{SEM}{model of class Rcpp_SEMCpp.}

\item{raw}{controls if the internal transformations of linr is used.}

\item{method}{optimizer method. See ?optim}

\item{control}{control optimizer. See ?optim}

\item{hessian}{Should the optimizer return the hessian?}
}
\description{
optimize a SEM
}
