% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-aCV4regularizedSEMWithCustomPenalty.R
\name{plot,aCV4regularizedSEMWithCustomPenalty,ANY-method}
\alias{plot,aCV4regularizedSEMWithCustomPenalty,ANY-method}
\title{plot}
\usage{
\S4method{plot}{aCV4regularizedSEMWithCustomPenalty,ANY}(x)
}
\arguments{
\item{x}{object of class aCV4regularizedSEMWithCustomPenalty}
}
\description{
plots the cross-validation fits
}
