% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-approximateInfluence.R
\name{plot,approximateInfluence,ANY-method}
\alias{plot,approximateInfluence,ANY-method}
\title{plot}
\usage{
\S4method{plot}{approximateInfluence,ANY}(x, alpha = NULL)
}
\arguments{
\item{object}{object of class approximateInfluence}
}
\description{
plot
}
