% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-approximateInfluence.R
\name{plot,approximateInfluenceFitIndices,ANY-method}
\alias{plot,approximateInfluenceFitIndices,ANY-method}
\title{plot}
\usage{
\S4method{plot}{approximateInfluenceFitIndices,ANY}(x, alpha = NULL, interactive = FALSE)
}
\arguments{
\item{x}{object of class approximateInfluenceFitIndices}

\item{alpha}{the alpha tuning parameter value for which the fit index should be plotted}

\item{interactive}{boolean: If set to TRUE, plotly will be used to make the plot interactive. This can provide more information into which person may influence the results substantially}
}
\description{
plot
}
