% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CV4regularizedSEMWithCustomPenalty.R
\name{plot,cv4regularizedSEMWithCustomPenalty,ANY-method}
\alias{plot,cv4regularizedSEMWithCustomPenalty,ANY-method}
\title{plot}
\usage{
\S4method{plot}{cv4regularizedSEMWithCustomPenalty,ANY}(x)
}
\arguments{
\item{x}{object of class cv4regularizedSEMWithCustomPenalty}
}
\description{
plots the cross-validation fits
}
