% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cvRegularizedSEM.R
\name{plot,cvRegularizedSEM,ANY-method}
\alias{plot,cvRegularizedSEM,ANY-method}
\title{plot}
\usage{
\S4method{plot}{cvRegularizedSEM,ANY}(x)
}
\arguments{
\item{x}{object of class regularizedSEM}
}
\description{
plots the regularized and unregularized parameters as well as the cross-validation fits for all levels of lambda
}
