% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cvRegularizedSEM.R
\name{plot,cvRegularizedSEM,missing-method}
\alias{plot,cvRegularizedSEM,missing-method}
\title{plots the cross-validation fits}
\usage{
\S4method{plot}{cvRegularizedSEM,missing}(x, y, ...)
}
\arguments{
\item{x}{object of class cvRegularizedSEM}

\item{y}{not used}

\item{...}{not used}
}
\description{
plots the cross-validation fits
}
