% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-gpRegularized.R
\name{plot,gpRegularized,ANY-method}
\alias{plot,gpRegularized,ANY-method}
\title{plot}
\usage{
\S4method{plot}{gpRegularized,ANY}(x, alpha = NULL, regularizedOnly = TRUE)
}
\arguments{
\item{x}{object of class gpRegularized}

\item{alpha}{numeric: value of alpha for which parameters should be returned. Must be present in object. Required if elastic net was used}

\item{regularizedOnly}{boolean: should only regularized parameters be plotted?``}
}
\description{
plots the regularized and unregularized parameters for all levels of lambda
}
