% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-gpRegularized.R
\name{plot,gpRegularized,missing-method}
\alias{plot,gpRegularized,missing-method}
\title{plots the regularized and unregularized parameters for all levels of lambda}
\usage{
\S4method{plot}{gpRegularized,missing}(x, y, ...)
}
\arguments{
\item{x}{object of class gpRegularized}

\item{y}{not used}

\item{...}{use regularizedOnly=FALSE to plot all parameters}
}
\value{
either an object of ggplot2 or of plotly
}
\description{
plots the regularized and unregularized parameters for all levels of lambda
}
