% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-gpRegularized.R
\docType{methods}
\name{plot-gpRegularized}
\alias{plot-gpRegularized}
\alias{plot,gpRegularized-method}
\title{plots the regularized and unregularized parameters for all levels of lambda}
\usage{
\S4method{plot}{gpRegularized,ANY}(x, regularizedOnly = TRUE)
}
\arguments{
\item{x}{object of class gpRegularized}

\item{regularizedOnly}{boolean: should only regularized parameters be plotted?``}
}
\description{
plots the regularized and unregularized parameters for all levels of lambda
}
