% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-stabSel.R
\name{plot,stabSel,missing-method}
\alias{plot,stabSel,missing-method}
\title{plots the regularized and unregularized parameters for all levels of the tuning parameters}
\usage{
\S4method{plot}{stabSel,missing}(x, y, ...)
}
\arguments{
\item{x}{object of class stabSel}

\item{y}{not used}

\item{...}{use regularizedOnly=FALSE to plot all parameters}
}
\value{
either an object of ggplot2 or of plotly
}
\description{
plots the regularized and unregularized parameters for all levels of the tuning parameters
}
