% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-misc.R
\docType{class}
\name{Rcpp_bfgsEnetMgSEM-class}
\alias{Rcpp_bfgsEnetMgSEM-class}
\title{Wrapper for C++ module. See ?lessSEM::bfgsEnetMgSEM}
\description{
Wrapper for C++ module. See ?lessSEM::bfgsEnetMgSEM
}
\keyword{internal}
