% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-misc.R
\docType{class}
\name{Rcpp_bfgsEnetSEM-class}
\alias{Rcpp_bfgsEnetSEM-class}
\title{Wrapper for C++ module. See ?lessSEM::bfgsEnetSEM}
\description{
Wrapper for C++ module. See ?lessSEM::bfgsEnetSEM
}
\keyword{internal}
