% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-misc.R
\docType{class}
\name{Rcpp_glmnetCappedL1MgSEM-class}
\alias{Rcpp_glmnetCappedL1MgSEM-class}
\title{Wrapper for C++ module. See ?lessSEM::glmnetCappedL1MgSEM}
\description{
Wrapper for C++ module. See ?lessSEM::glmnetCappedL1MgSEM
}
\keyword{internal}
