% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-misc.R
\docType{class}
\name{Rcpp_glmnetCappedL1SEM-class}
\alias{Rcpp_glmnetCappedL1SEM-class}
\title{Wrapper for C++ module. See ?lessSEM:::glmnetCappedL1SEM}
\description{
Wrapper for C++ module. See ?lessSEM:::glmnetCappedL1SEM
}
\keyword{internal}
