% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-misc.R
\docType{class}
\name{Rcpp_glmnetLspSEM-class}
\alias{Rcpp_glmnetLspSEM-class}
\title{Wrapper for C++ module. See ?lessSEM:::glmnetLspSEM}
\description{
Wrapper for C++ module. See ?lessSEM:::glmnetLspSEM
}
\keyword{internal}
