% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-misc.R
\docType{class}
\name{Rcpp_glmnetMcpMgSEM-class}
\alias{Rcpp_glmnetMcpMgSEM-class}
\title{Wrapper for C++ module. See ?lessSEM::glmnetMcpMgSEM}
\description{
Wrapper for C++ module. See ?lessSEM::glmnetMcpMgSEM
}
\keyword{internal}
