% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-misc.R
\docType{class}
\name{Rcpp_glmnetMcpSEM-class}
\alias{Rcpp_glmnetMcpSEM-class}
\title{Wrapper for C++ module. See ?lessSEM:::glmnetMcpSEM}
\description{
Wrapper for C++ module. See ?lessSEM:::glmnetMcpSEM
}
\keyword{internal}
