% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-misc.R
\docType{class}
\name{Rcpp_glmnetScadMgSEM-class}
\alias{Rcpp_glmnetScadMgSEM-class}
\title{Wrapper for C++ module. See ?lessSEM::glmnetScadMgSEM}
\description{
Wrapper for C++ module. See ?lessSEM::glmnetScadMgSEM
}
\keyword{internal}
