% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-misc.R
\docType{class}
\name{Rcpp_glmnetScadSEM-class}
\alias{Rcpp_glmnetScadSEM-class}
\title{Wrapper for C++ module. See ?lessSEM:::glmnetScadSEM}
\description{
Wrapper for C++ module. See ?lessSEM:::glmnetScadSEM
}
\keyword{internal}
