% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-misc.R
\docType{class}
\name{Rcpp_istaCappedL1GeneralPurpose-class}
\alias{Rcpp_istaCappedL1GeneralPurpose-class}
\title{Wrapper for C++ module. See ?lessSEM::istaCappedL1GeneralPurpose}
\description{
Wrapper for C++ module. See ?lessSEM::istaCappedL1GeneralPurpose
}
\keyword{internal}
