% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-misc.R
\name{Rcpp_istaCappedL1GeneralPurpose}
\alias{Rcpp_istaCappedL1GeneralPurpose}
\title{Wrapper for C++ module. See ?lessSEM::istaCappedL1GeneralPurpose}
\description{
Wrapper for C++ module. See ?lessSEM::istaCappedL1GeneralPurpose
}
