% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-misc.R
\docType{class}
\name{Rcpp_istaCappedL1GeneralPurposeCpp-class}
\alias{Rcpp_istaCappedL1GeneralPurposeCpp-class}
\title{Wrapper for C++ module. See ?lessSEM::istaCappedL1GeneralPurposeCpp}
\description{
Wrapper for C++ module. See ?lessSEM::istaCappedL1GeneralPurposeCpp
}
\keyword{internal}
