% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-misc.R
\docType{class}
\name{Rcpp_istaCappedL1MgSEM-class}
\alias{Rcpp_istaCappedL1MgSEM-class}
\title{Wrapper for C++ module. See ?lessSEM::istaCappedL1MgSEM}
\description{
Wrapper for C++ module. See ?lessSEM::istaCappedL1MgSEM
}
\keyword{internal}
