% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-misc.R
\docType{class}
\name{Rcpp_istaCappedL1SEM-class}
\alias{Rcpp_istaCappedL1SEM-class}
\title{Wrapper for C++ module. See ?lessSEM::istaCappedL1SEM}
\description{
Wrapper for C++ module. See ?lessSEM::istaCappedL1SEM
}
\keyword{internal}
