% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-misc.R
\docType{class}
\name{Rcpp_istaEnetGeneralPurpose}
\alias{Rcpp_istaEnetGeneralPurpose}
\title{Wrapper for C++ module. See ?lessSEM::istaEnetGeneralPurpose}
\description{
Wrapper for C++ module. See ?lessSEM::istaEnetGeneralPurpose
}
