% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-misc.R
\docType{class}
\name{Rcpp_istaEnetGeneralPurposeCpp-class}
\alias{Rcpp_istaEnetGeneralPurposeCpp-class}
\title{Wrapper for C++ module. See ?lessSEM::istaEnetGeneralPurposeCpp}
\description{
Wrapper for C++ module. See ?lessSEM::istaEnetGeneralPurposeCpp
}
\keyword{internal}
