% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-misc.R
\docType{class}
\name{Rcpp_istaEnetMgSEM-class}
\alias{Rcpp_istaEnetMgSEM-class}
\title{Wrapper for C++ module. See ?lessSEM::istaEnetMgSEM}
\description{
Wrapper for C++ module. See ?lessSEM::istaEnetMgSEM

Wrapper for C++ module. See ?lessSEM::istaEnetMgSEM
}
\keyword{internal}
