% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-misc.R
\docType{class}
\name{Rcpp_istaLspGeneralPurpose-class}
\alias{Rcpp_istaLspGeneralPurpose-class}
\title{Wrapper for C++ module. See ?lessSEM::istaLspGeneralPurpose}
\description{
Wrapper for C++ module. See ?lessSEM::istaLspGeneralPurpose
}
\keyword{internal}
