% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-misc.R
\docType{class}
\name{Rcpp_istaLspGeneralPurposeCpp-class}
\alias{Rcpp_istaLspGeneralPurposeCpp-class}
\title{Wrapper for C++ module. See ?lessSEM::istaLspGeneralPurposeCpp}
\description{
Wrapper for C++ module. See ?lessSEM::istaLspGeneralPurposeCpp
}
\keyword{internal}
