% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-misc.R
\name{Rcpp_istaLspGeneralPurposeCpp}
\alias{Rcpp_istaLspGeneralPurposeCpp}
\title{Wrapper for C++ module. See ?lessSEM::istaLspGeneralPurposeCpp}
\description{
Wrapper for C++ module. See ?lessSEM::istaLspGeneralPurposeCpp
}
