% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-misc.R
\docType{class}
\name{Rcpp_istaLSPMgSEM-class}
\alias{Rcpp_istaLSPMgSEM-class}
\title{Wrapper for C++ module. See ?lessSEM::istaLSPMgSEM}
\description{
Wrapper for C++ module. See ?lessSEM::istaLSPMgSEM
}
\keyword{internal}
