% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-misc.R
\docType{class}
\name{Rcpp_istaLSPSEM-class}
\alias{Rcpp_istaLSPSEM-class}
\title{Wrapper for C++ module. See ?lessSEM::istaLSPSEM}
\description{
Wrapper for C++ module. See ?lessSEM::istaLSPSEM
}
\keyword{internal}
