% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-misc.R
\docType{class}
\name{Rcpp_istaMcpGeneralPurpose-class}
\alias{Rcpp_istaMcpGeneralPurpose-class}
\title{Wrapper for C++ module. See ?lessSEM::istaMcpGeneralPurpose}
\description{
Wrapper for C++ module. See ?lessSEM::istaMcpGeneralPurpose
}
\keyword{internal}
