% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-misc.R
\docType{class}
\name{Rcpp_istaMcpGeneralPurposeCpp}
\alias{Rcpp_istaMcpGeneralPurposeCpp}
\title{Wrapper for C++ module. See ?lessSEM::istaMcpGeneralPurposeCpp}
\description{
Wrapper for C++ module. See ?lessSEM::istaMcpGeneralPurposeCpp
}
