% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-misc.R
\docType{class}
\name{Rcpp_istaMcpMgSEM-class}
\alias{Rcpp_istaMcpMgSEM-class}
\title{Wrapper for C++ module. See ?lessSEM::istaMcpMgSEM}
\description{
Wrapper for C++ module. See ?lessSEM::istaMcpMgSEM
}
\keyword{internal}
