% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-misc.R
\docType{class}
\name{Rcpp_istaMixedPenaltymgSEM-class}
\alias{Rcpp_istaMixedPenaltymgSEM-class}
\title{Wrapper for C++ module. See ?lessSEM::istaMixedPenaltymgSEM}
\description{
Wrapper for C++ module. See ?lessSEM::istaMixedPenaltymgSEM
}
\keyword{internal}
