% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-misc.R
\docType{class}
\name{Rcpp_istaMixedPenaltySEM-class}
\alias{Rcpp_istaMixedPenaltySEM-class}
\title{Wrapper for C++ module. See ?lessSEM::istaMixedPenaltySEM}
\description{
Wrapper for C++ module. See ?lessSEM::istaMixedPenaltySEM
}
\keyword{internal}
