% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-misc.R
\docType{class}
\name{Rcpp_istaScadGeneralPurpose}
\alias{Rcpp_istaScadGeneralPurpose}
\title{Wrapper for C++ module. See ?lessSEM::istaScadGeneralPurpose}
\description{
Wrapper for C++ module. See ?lessSEM::istaScadGeneralPurpose
}
