% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-misc.R
\docType{class}
\name{Rcpp_istaScadGeneralPurposeCpp-class}
\alias{Rcpp_istaScadGeneralPurposeCpp-class}
\title{Wrapper for C++ module. See ?lessSEM::istaScadGeneralPurposeCpp}
\description{
Wrapper for C++ module. See ?lessSEM::istaScadGeneralPurposeCpp
}
\keyword{internal}
