% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-misc.R
\name{Rcpp_istaScadGeneralPurposeCpp}
\alias{Rcpp_istaScadGeneralPurposeCpp}
\title{Wrapper for C++ module. See ?lessSEM::istaScadGeneralPurposeCpp}
\description{
Wrapper for C++ module. See ?lessSEM::istaScadGeneralPurposeCpp
}
