% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-misc.R
\docType{class}
\name{Rcpp_istaScadMgSEM-class}
\alias{Rcpp_istaScadMgSEM-class}
\title{Wrapper for C++ module. See ?lessSEM::istaScadMgSEM}
\description{
Wrapper for C++ module. See ?lessSEM::istaScadMgSEM
}
\keyword{internal}
