% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-misc.R
\docType{class}
\name{Rcpp_istaScadSEM-class}
\alias{Rcpp_istaScadSEM-class}
\title{Wrapper for C++ module. See ?lessSEM::istaScadSEM}
\description{
Wrapper for C++ module. See ?lessSEM::istaScadSEM
}
\keyword{internal}
