% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaanHelperFunctions.R
\name{regsem2LavaanParameters}
\alias{regsem2LavaanParameters}
\title{regsem2LavaanParameters}
\usage{
regsem2LavaanParameters(regsemModel, lavaanModel)
}
\arguments{
\item{regsemModel}{model of class regsem}

\item{lavaanModel}{model of class lavaan}
}
\description{
helper function: regsem and lavaan use slightly different parameter labels. This function
can be used to translate the parameter labels of a cv_regsem object to lavaan labels
}
