% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-regularizedSEMMixedPenalty.R
\docType{class}
\name{regularizedSEMMixedPenalty-class}
\alias{regularizedSEMMixedPenalty-class}
\title{Class for regularized SEM}
\description{
Class for regularized SEM
}
\section{Slots}{

\describe{
\item{\code{penalty}}{penalty used (e.g., "lasso")}

\item{\code{tuningParameterConfigurations}}{list with settings for the lambda, theta,
and alpha tuning parameters.}

\item{\code{parameters}}{data.frame with parameter estimates}

\item{\code{fits}}{data.frame with all fit results}

\item{\code{parameterLabels}}{character vector with names of all parameters}

\item{\code{weights}}{vector with weights given to each of the parameters in the penalty}

\item{\code{regularized}}{character vector with names of regularized parameters}

\item{\code{transformations}}{if the model has transformations, the transformed parameters
are returned}

\item{\code{internalOptimization}}{list of elements used internally}

\item{\code{inputArguments}}{list with elements passed by the user to the general}
}}

