% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-regularizedSEMWithCustomPenalty.R
\docType{class}
\name{regularizedSEMWithCustomPenalty-class}
\alias{regularizedSEMWithCustomPenalty-class}
\title{Class for regularized SEM using Rsolnp}
\description{
Class for regularized SEM using Rsolnp
}
\section{Slots}{

\describe{
\item{\code{parameters}}{data.frame with parameter estimates}

\item{\code{fits}}{data.frame with all fit results}

\item{\code{parameterLabels}}{character vector with names of all parameters}

\item{\code{internalOptimization}}{list of elements used internally}

\item{\code{inputArguments}}{list with elements passed by the user to the general}

\item{\code{notes}}{internal notes that have come up when fitting the model}
}}

\keyword{internal}
