% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularizeSEM.R
\name{regularizeSEM}
\alias{regularizeSEM}
\title{regularizeSEM}
\usage{
regularizeSEM(
  lavaanModel,
  regularizedParameterLabels,
  penalty,
  lambdas = NULL,
  nLambdas = NULL,
  alphas = NULL,
  adaptiveLassoWeights = NULL,
  raw = TRUE,
  control = aCV4SEM::controlGLMNET()
)
}
\arguments{
\item{lavaanModel}{model of class lavaan}

\item{regularizedParameterLabels}{labels of regularized parameters}

\item{penalty}{which penalty should be used? Available are "ridge", "lasso", "adaptiveLasso", and "elasticNet"}

\item{lambdas}{vector with lambda values. Higher values = higher penalty}

\item{nLambdas}{if penalty == "lasso" or penalty == "adaptiveLasso", one can specify the number of lambda values to use. In this case, set lambdas = NULL.}

\item{alphas}{0<alpha<1. only required for elastic net. Controls the weight of ridge and lasso terms. alpha = 1 is lasso, alpha = 0 ridge}

\item{adaptiveLassoWeights}{vector with weights for adaptive LASSO. Set to NULL if not using adaptive LASSO}

\item{raw}{controls if the internal transformations of aCV4SEM is used.}

\item{control}{option to set parameters of the GLMNET optimizer}
}
\description{
optimize a SEM with regularized penalty function.
}
