% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularizeSEMWithCustomPenalty.R
\name{regularizeSEMWithCustomPenalty}
\alias{regularizeSEMWithCustomPenalty}
\title{regularizeSEMWithCustomPenalty}
\usage{
regularizeSEMWithCustomPenalty(
  lavaanModel,
  individualPenaltyFunction,
  individualPenaltyFunctionGradient = NULL,
  individualPenaltyFunctionHessian = NULL,
  tuningParameters,
  penaltyFunctionArguments,
  control = aCV4SEM::controlQuasiNewtonBFGS()
)
}
\arguments{
\item{lavaanModel}{model of class lavaan}

\item{individualPenaltyFunction}{penalty function which takes the current parameter values as first argument, the tuning parameters as second, and the penaltyFunctionArguments as third argument and 
returns a single value - the value of the penalty function for a single person. If the true penalty function is non-differentiable (e.g., lasso) a smooth
approximation of this function should be provided.}

\item{individualPenaltyFunctionGradient}{gradients of the penalty function. Function should take the current parameter values as first argument, the tuning parameters as second, and the penaltyFunctionArguments as third argument and 
return a vector of the same length as parameters. If the true penalty function is non-differentiable (e.g., lasso) a smooth
approximation of this function should be provided. If set to NULL, numDeriv will be used.}

\item{individualPenaltyFunctionHessian}{Hessian of the penalty function. Function should take the current parameter values as first argument, the tuning parameters as second, and the penaltyFunctionArguments as third argument and 
return a matrix with (length as parameters)^2 number of elements. If the true penalty function is non-differentiable (e.g., lasso) a smooth
approximation of this function should be provided. If set to NULL, numDeriv will be used.}

\item{tuningParameters}{data.frame with tuning parameter values. Important: The function will iterate over the rows of these tuning parameters and pass them to your penalty function}

\item{penaltyFunctionArguments}{arguments passed to individualPenaltyFunction, individualPenaltyFunctionGradient, and individualPenaltyFunctionHessian}

\item{control}{option to set parameters of the optimizer}
}
\description{
optimize a SEM with custom penalty functions.
}
