% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penaltyFunctions.R
\name{ridgeHessian}
\alias{ridgeHessian}
\title{ridgeGradient}
\usage{
ridgeHessian(parameters, penaltyFunctionArguments)
}
\arguments{
\item{parameters}{vector with labeled parameter values}

\item{penaltyFunctionArguments}{list with fields lambda (tuning parameter value) and regularizedParameterLabels (labels of regularized parameters)}
}
\description{
ridge Hessian function
}
