% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scadPenalty_C}
\alias{scadPenalty_C}
\title{scadPenalty_C}
\usage{
scadPenalty_C(par, lambda_p, theta)
}
\arguments{
\item{par}{single parameter value}

\item{lambda_p}{lambda value for this parameter}

\item{theta}{theta value for this parameter}
}
\value{
penalty value
}
\description{
scadPenalty_C
}
