% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unregularizedSEM.R
\name{SEM}
\alias{SEM}
\title{SEM}
\usage{
SEM(lavaanSyntax, rawData, optimizer = "lbfgs")
}
\arguments{
\item{lavaanSyntax}{syntax used to generate a model for lavaan}

\item{rawData}{matrix with data used for the model}

\item{optimizer}{optimizer to use. SEM relies on optimizers from ensmallen}
}
\value{
object of class SEMCpp
}
\description{
Fit an unregularized SEM. The function builds on the brilliant lavaan package which 
simplifies setting up models considerably. The main objective of SEM is to
test the ensmallen optimizers.
}
\examples{
library(lavaan)
library(aCV4SEM)
set.seed(123)
modelSyntax <- ' 
  # latent variable definitions
     ind60 =~ x1 + x2 + x3
     dem60 =~ y1 + a*y2 + b*y3 + c*y4
     dem65 =~ y5 + a*y6 + b*y7 + c*y8

  # regressions
    dem60 ~ ind60
    dem65 ~ ind60 + dem60

  # residual correlations
    y1 ~~ y5
    y2 ~~ y4 
    y3 ~~ y7
    y4 ~~ y8
    y6 ~~ y8
'

SEM <- aCV4SEM:::SEM(lavaanSyntax = modelSyntax,
                     rawData = PoliticalDemocracy)
}
