% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SEMCpp}
\alias{SEMCpp}
\title{SEMCpp class}
\description{
internal SEM representation
}
\section{Fields}{

\describe{
\item{\code{new}}{Creates a new SEMCpp.}

\item{\code{A}}{Matrix with directed effects}

\item{\code{S}}{Matrix with undirected paths}

\item{\code{F}}{Filter matrix to separate latent and observed variables}

\item{\code{m}}{Vector with means of observed and latent variables}

\item{\code{impliedCovariance}}{implied covariance matrix}

\item{\code{impliedMeans}}{implied means vector}

\item{\code{m2LL}}{-2 log-likelihood}

\item{\code{rawData}}{raw data set}

\item{\code{manifestNames}}{names of manifest variables}

\item{\code{wasFit}}{names of manifest variables}

\item{\code{setMatrix}}{Fills the elements of a model matrix. Expects a char 
(A, S, or F), and a matrix with values}

\item{\code{setVector}}{Fills the elements of a model vector. 
Expects a char (m), and a vector with values}

\item{\code{initializeParameters}}{Initializes the parameters of the model. 
Expects StringVector with labels, StringVector with location, 
uvec with rows, uvec with columns, vec with values, and vec with rawValues}

\item{\code{setParameters}}{Changes the parameters of a model. Expects StringVector 
with labels, vec with values, and boolean indicating if the values are raw (TRUE) 
or transformed (FALSE).}

\item{\code{addDerivativeElement}}{Add an element to the internal derivative structure. 
string label, string with location, and boolean indicating if it is a variance, matrix with positions.}

\item{\code{addRawData}}{Adds a raw data set. Expects matrix and vector with labels of manifest variables.}

\item{\code{addSubset}}{Adds a subset to the data. Expects the sample size N of the subset, 
the number of observed variables without missings, uvec with indices of notMissing values, 
matrix with covariance, colvec with means, raw data without missings.}

\item{\code{implied}}{Computes implied means and covariance matrix}

\item{\code{fit}}{Fits the model. Returns -2 log likelihood}

\item{\code{getParameters}}{Returns a data frame with model parameters.}

\item{\code{getParameterLabels}}{Returns a vector with unique parameter labels as used internally.}

\item{\code{getGradients}}{Returns a matrix with scores.}

\item{\code{getScores}}{Returns a matrix with scores.}

\item{\code{getHessian}}{Returns the hessian of the model. Expects the labels of the 
parameters and the values of the parameters as well as a boolean indicating if 
these are raw. Finally, a double (eps) controls the precision of the approximation.}
}}

