% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-gpRegularized.R
\name{show,gpRegularized-method}
\alias{show,gpRegularized-method}
\title{show}
\usage{
\S4method{show}{gpRegularized}(object)
}
\arguments{
\item{object}{object of class gpRegularized}
}
\value{
No return value, just prints estimates
}
\description{
show
}
