% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-coef.R
\name{show,lessSEMCoef-method}
\alias{show,lessSEMCoef-method}
\title{show}
\usage{
\S4method{show}{lessSEMCoef}(object)
}
\arguments{
\item{object}{object of class lessSEMCoef}
}
\value{
No return value, just prints estimates
}
\description{
show
}
