% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-logLikelihood.R
\name{show,logLikelihood-method}
\alias{show,logLikelihood-method}
\title{show}
\usage{
\S4method{show}{logLikelihood}(object)
}
\arguments{
\item{object}{object of class logLikelihood}
}
\value{
No return value, just prints estimates
}
\description{
show
}
