% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Rcpp_mgSEM.R
\name{show,Rcpp_mgSEM-method}
\alias{show,Rcpp_mgSEM-method}
\title{show}
\usage{
\S4method{show}{Rcpp_mgSEM}(object)
}
\arguments{
\item{object}{object of class Rcpp_mgSEM}
}
\value{
No return value, just prints estimates
}
\description{
show
}
