% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Rcpp_SEMCpp.R
\name{show,Rcpp_SEMCpp-method}
\alias{show,Rcpp_SEMCpp-method}
\title{show}
\usage{
\S4method{show}{Rcpp_SEMCpp}(object)
}
\arguments{
\item{object}{object of class Rcpp_SEMCpp}
}
\value{
No return value, just prints estimates
}
\description{
show
}
