% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-regularizedSEM.R
\name{show,regularizedSEM-method}
\alias{show,regularizedSEM-method}
\title{show}
\usage{
\S4method{show}{regularizedSEM}(object)
}
\arguments{
\item{object}{object of class regularizedSEM}
}
\description{
show
}
