% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-regularizedSEMMixedPenalty.R
\name{show,regularizedSEMMixedPenalty-method}
\alias{show,regularizedSEMMixedPenalty-method}
\title{show}
\usage{
\S4method{show}{regularizedSEMMixedPenalty}(object)
}
\arguments{
\item{object}{object of class regularizedSEM}
}
\value{
No return value, just prints estimates
}
\description{
show
}
