% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-stabSel.R
\name{show,stabSel-method}
\alias{show,stabSel-method}
\title{show}
\usage{
\S4method{show}{stabSel}(object)
}
\arguments{
\item{object}{object of class stabSel}
}
\value{
No return value, just prints estimates
}
\description{
show
}
