% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penaltyFunctions.R
\name{smoothAdaptiveLASSOGradient}
\alias{smoothAdaptiveLASSOGradient}
\title{smoothAdaptiveLASSOGradient}
\usage{
smoothAdaptiveLASSOGradient(parameters, penaltyFunctionArguments)
}
\arguments{
\item{parameters}{vector with labeled parameter values}

\item{penaltyFunctionArguments}{list with fields lambdas (vector with one tuning parameter value for each parameter), regularizedParameterLabels (labels of regularized parameters), and eps (controls the smooth approximation of non-differential penalty functions (e.g., lasso, adaptive lasso, or elastic net). Smaller values result in closer approximation, but may also cause larger issues in optimization.)}
}
\description{
smoothed version of non-differentiable adaptive LASSO gradient
}
