% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothOptimization.R
\name{smoothLasso}
\alias{smoothLasso}
\title{smoothLasso}
\usage{
smoothLasso(
  lavaanModel,
  regularized,
  lambdas = NULL,
  nLambdas = NULL,
  epsilon,
  control = controlBFGS()
)
}
\arguments{
\item{lavaanModel}{model of class lavaan}

\item{regularized}{vector with names of parameters which are to be regularized.
If you are unsure what these parameters are called, use
getLavaanParameters(model) with your lavaan model object}

\item{lambdas}{numeric vector: values for the tuning parameter lambda}

\item{nLambdas}{alternative to lambda: If alpha = 1, linr can automatically
compute the first lambda value which sets all regularized parameters to zero.
It will then generate nLambda values between 0 and the computed lambda.}

\item{epsilon}{epsilon > 0; controls the smoothness of the approximation. Larger values = smoother}

\item{control}{used to control the optimizer. This element is generated with
the controlBFGS function. See ?controlBFGS for more details.}
}
\description{
This function allows for regularization of models built in lavaan with the
smoothed lasso penalty. The returned object is an S4 class; its elements can be accessed
with the "@" operator (see examples). We don't recommend using this function.
Use lasso() instead.
}
\details{
For more details, see:
\enumerate{
\item Lee, S.-I., Lee, H., Abbeel, P., & Ng, A. Y. (2006).
Efficient L1 Regularized Logistic Regression. Proceedings of the
Twenty-First National Conference on Artificial Intelligence (AAAI-06), 401–408.
\item Jacobucci, R., Grimm, K. J., & McArdle, J. J. (2016).
Regularized Structural Equation Modeling. Structural Equation Modeling:
A Multidisciplinary Journal, 23(4), 555–566. https://doi.org/10.1080/10705511.2016.1154793
}
}
