% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-stabSel.R
\docType{class}
\name{stabSel-class}
\alias{stabSel-class}
\title{Class for stability selection}
\description{
Class for stability selection
}
\section{Slots}{

\describe{
\item{\code{tuningParameters}}{tuning parameters}

\item{\code{estimates}}{parameter estimates}

\item{\code{transformations}}{transformations of parameters}
}}

