% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-aCV4regularizedSEMWithCustomPenalty.R
\name{summary,aCV4regularizedSEMWithCustomPenalty-method}
\alias{summary,aCV4regularizedSEMWithCustomPenalty-method}
\title{summary}
\usage{
\S4method{summary}{aCV4regularizedSEMWithCustomPenalty}(object)
}
\description{
summary
}
