% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CV4regularizedSEMWithCustomPenalty.R
\name{summary,CV4regularizedSEMWithCustomPenalty-method}
\alias{summary,CV4regularizedSEMWithCustomPenalty-method}
\title{summary}
\usage{
\S4method{summary}{CV4regularizedSEMWithCustomPenalty}(object)
}
\description{
summary
}
