% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cvRegularizedSEM.R
\name{summary,cvRegularizedSEM-method}
\alias{summary,cvRegularizedSEM-method}
\title{summary method for objects of class \code{cvRegularizedSEM}.}
\usage{
\S4method{summary}{cvRegularizedSEM}(object, ...)
}
\arguments{
\item{object}{object of class cvRegularizedSEM}

\item{...}{not used}
}
\value{
No return value, just prints estimates
}
\description{
summary method for objects of class \code{cvRegularizedSEM}.
}
