% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cvRegularizedSEM.R
\docType{methods}
\name{summary-cvRegularizedSEM}
\alias{summary-cvRegularizedSEM}
\alias{summary,cvRegularizedSEM-method}
\title{summary method for objects of class \code{cvRegularizedSEM}.}
\usage{
\S4method{summary}{cvRegularizedSEM}(object)
}
\arguments{
\item{object}{object of class cvRegularizedSEM}
}
\description{
summary method for objects of class \code{cvRegularizedSEM}.
}
