% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-gpRegularized.R
\name{summary,gpRegularized-method}
\alias{summary,gpRegularized-method}
\title{summary}
\usage{
\S4method{summary}{gpRegularized}(object)
}
\arguments{
\item{object}{object of class gpRegularized}
}
\description{
summary
}
