% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-regularizedSEMMixedPenalty.R
\name{summary,regularizedSEMMixedPenalty-method}
\alias{summary,regularizedSEMMixedPenalty-method}
\title{summary}
\usage{
\S4method{summary}{regularizedSEMMixedPenalty}(object)
}
\arguments{
\item{object}{object of class regularizedSEMMixedPenalty}
}
\description{
summary
}
