% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-regularizedSEMWithCustomPenalty.R
\name{summary,regularizedSEMWithCustomPenalty-method}
\alias{summary,regularizedSEMWithCustomPenalty-method}
\title{summary}
\usage{
\S4method{summary}{regularizedSEMWithCustomPenalty}(object, ...)
}
\arguments{
\item{object}{object of class regularizedSEMWithCustomPenalty}

\item{...}{not used}
}
\value{
No return value, just prints estimates
}
\description{
summary
}
\keyword{internal}
