% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizeSEM.R
\name{wideResults}
\alias{wideResults}
\title{wideResults}
\usage{
wideResults(regularizedSEM)
}
\arguments{
\item{regularizedSEM}{object of class regularizedSEM from regularizeSEM() function}
}
\value{
matrix with parameters in wide format
}
\description{
The results of regularizeSEM are returned in tidy format. If you prefer wide format, this function will return just that
}
