% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aCV4model.R
\name{aCV4regularizedSEM}
\alias{aCV4regularizedSEM}
\title{aCV4regularizedSEM}
\usage{
aCV4regularizedSEM(regularizedSEM, k, recomputeHessian = TRUE)
}
\arguments{
\item{regularizedSEM}{model of class regularizedSEM}

\item{k}{the number of cross-validation folds. We recommend leave-one-out cross-validation; i.e. set k to the number of persons in the data set}

\item{recomputeHessian}{if set to false, the Hessians from the quasi newton optimization with GLMNET will be used. Otherwise the Hessian will be recomputed.}
}
\description{
approximate cross-validation for models of class regularizedSEM. These models can be fit with regularizedSEM() (see ?regularizedSEM)
in this package.
}
