% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aCV4model.R
\name{aCV4regularizedSEMWithCustomPenalty}
\alias{aCV4regularizedSEMWithCustomPenalty}
\title{aCV4regularizedSEMWithCustomPenalty}
\usage{
aCV4regularizedSEMWithCustomPenalty(
  regularizedSEMWithCustomPenalty,
  k,
  returnSubsetParameters = FALSE,
  recomputeHessian = TRUE
)
}
\arguments{
\item{k}{the number of cross-validation folds. We recommend leave-one-out cross-validation; i.e. set k to the number of persons in the data set. Alternatively, 
a matrix with pre-defined subsets can be passed to the function. See ?linr::aCV4regularizedSEM for an example}

\item{returnSubsetParameters}{if set to TRUE, the parameter estimates of the individual cross-validation training sets will be returned}

\item{recomputeHessian}{if set to FALSE, the Hessians from the quasi newton optimization with BFGS will be used. Otherwise the Hessian will be recomputed. We currently recommend setting recomputeHessian to TRUE}

\item{regularizedSEMWithCustomPenalty.}{model of class regularizedSEMWithCustomPenalty.}
}
\description{
approximate cross-validation for models of class regularizedSEMWithCustomPenalty. These models can be fit with regularizedSEMWithCustomPenalty.() (see ?regularizedSEMWithCustomPenalty.)
in this package.
}
