% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-gpRegularized.R
\name{AIC,gpRegularized-method}
\alias{AIC,gpRegularized-method}
\title{AIC}
\usage{
\S4method{AIC}{gpRegularized}(object, ..., k = 2)
}
\arguments{
\item{object}{object of class gpRegularized}

\item{...}{not used}

\item{k}{multiplier for number of parameters}
}
\value{
data frame with fit values, appended with AIC
}
\description{
returns the AIC
}
