% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-regularizedSEM.R
\name{AIC,regularizedSEM-method}
\alias{AIC,regularizedSEM-method}
\title{AIC}
\usage{
\S4method{AIC}{regularizedSEM}(object, ..., k = 2)
}
\arguments{
\item{object}{object of class regularizedSEM}

\item{...}{not used}

\item{k}{multiplier for number of parameters}
}
\value{
AIC values
}
\description{
returns the AIC
}
