% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-regularizedSEMMixedPenalty.R
\name{AIC,regularizedSEMMixedPenalty-method}
\alias{AIC,regularizedSEMMixedPenalty-method}
\title{AIC}
\usage{
\S4method{AIC}{regularizedSEMMixedPenalty}(object, ..., k = 2)
}
\arguments{
\item{object}{object of class regularizedSEMMixedPenalty}

\item{...}{not used}

\item{k}{multiplier for number of parameters}
}
\value{
AIC values
}
\description{
returns the AIC
}
