% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-regularizedSEMWithCustomPenalty.R
\name{AIC,regularizedSEMWithCustomPenalty-method}
\alias{AIC,regularizedSEMWithCustomPenalty-method}
\title{AIC}
\usage{
\S4method{AIC}{regularizedSEMWithCustomPenalty}(object, penalizedParameterLabels, zeroThreshold)
}
\arguments{
\item{object}{object of class regularizedSEMWithCustomPenalty}

\item{penalizedParameterLabels}{vector with labels of penalized parameters}

\item{zeroThreshold}{penalized parameters below this threshold will be counted as zeroed}
}
\description{
returns the AIC
}
