% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-regularizedSEMWithCustomPenalty.R
\name{AIC,regularizedSEMWithCustomPenalty-method}
\alias{AIC,regularizedSEMWithCustomPenalty-method}
\title{AIC}
\usage{
\S4method{AIC}{regularizedSEMWithCustomPenalty}(object, ..., k = 2)
}
\arguments{
\item{object}{object of class regularizedSEMWithCustomPenalty}

\item{...}{Expects penalizedParameterLabels and zeroThreshold.
penalizedParameterLabels: vector with labels of penalized parameters.
zeroThreshold: penalized parameters below this threshold will be counted as zeroed.}

\item{k}{multiplier for number of parameters}
}
\value{
AIC values
}
\description{
returns the AIC. Expects penalizedParameterLabels and zeroThreshold
}
\keyword{internal}
