% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-regularizedSEMWithCustomPenalty.R
\name{BIC,regularizedSEMWithCustomPenalty-method}
\alias{BIC,regularizedSEMWithCustomPenalty-method}
\title{BIC}
\usage{
\S4method{BIC}{regularizedSEMWithCustomPenalty}(object, ...)
}
\arguments{
\item{object}{object of class regularizedSEMWithCustomPenalty}

\item{...}{Expects penalizedParameterLabels and zeroThreshold.
penalizedParameterLabels: vector with labels of penalized parameters.
zeroThreshold: penalized parameters below this threshold will be counted as zeroed.}
}
\value{
BIC values
}
\description{
returns the BIC
}
\keyword{internal}
