% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CV4regularizedSEMWithCustomPenalty.R
\name{coef,cv4regularizedSEMWithCustomPenalty-method}
\alias{coef,cv4regularizedSEMWithCustomPenalty-method}
\title{coef}
\usage{
\S4method{coef}{cv4regularizedSEMWithCustomPenalty}(object)
}
\arguments{
\item{object}{object of class regularizedSEM}
}
\description{
Returns the parameter estimates of an cv4regularizedSEMWithCustomPenalty
}
