% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-gpRegularized.R
\name{coef,gpRegularized-method}
\alias{coef,gpRegularized-method}
\title{coef}
\usage{
\S4method{coef}{gpRegularized}(object, ...)
}
\arguments{
\item{object}{object of class gpRegularized}

\item{...}{criterion can be one of: "AIC", "BIC". If set to NULL, all parameters will be returned}
}
\value{
parameter estimates
}
\description{
Returns the parameter estimates of a gpRegularized
}
