% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Rcpp_SEMCpp.R
\name{coef,Rcpp_SEMCpp-method}
\alias{coef,Rcpp_SEMCpp-method}
\title{coef}
\usage{
\S4method{coef}{Rcpp_SEMCpp}(object, ...)
}
\arguments{
\item{object}{object of class Rcpp_SEMCpp}

\item{...}{not used}
}
\value{
all coefficients of the model in transformed form
}
\description{
coef
}
