% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-regularizedSEM.R
\name{coef,regularizedSEM-method}
\alias{coef,regularizedSEM-method}
\title{coef}
\usage{
\S4method{coef}{regularizedSEM}(object, ...)
}
\arguments{
\item{object}{object of class regularizedSEM}

\item{...}{criterion can be one of the ones returned by fitIndices. If set to NULL, all parameters will be returned}
}
\value{
parameters of the model as data.frame
}
\description{
Returns the parameter estimates of a regularizedSEM
}
