% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeImpliedCovariance}
\alias{computeImpliedCovariance}
\title{computeImpliedCovariance}
\usage{
computeImpliedCovariance(Fmatrix, Amatrix, Smatrix)
}
\arguments{
\item{Fmatrix}{filter matrix}

\item{Amatrix}{matrix with directed effects}

\item{Smatrix}{matrix with undirected effects}
}
\value{
matrix with implied covariances
}
\description{
computes the implied covariance matrix of a SEM using RAM notation
}
