% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeImpliedMeans}
\alias{computeImpliedMeans}
\title{computeImpliedMeans}
\usage{
computeImpliedMeans(Fmatrix, Amatrix, Mvector)
}
\arguments{
\item{Fmatrix}{filter matrix}

\item{Amatrix}{matrix with directed effects}

\item{Mvector}{vector with means}
}
\value{
matrix with implied means
}
\description{
computes the implied means vector of a SEM using RAM notation
}
