% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{.compileTransformations}
\alias{.compileTransformations}
\title{.compileTransformations}
\usage{
.compileTransformations(syntax)
}
\arguments{
\item{syntax}{string with user defined transformations}
}
\value{
list with parameter names and two Rcpp functions: (1) the transformation function and
(2) a function to create a pointer to the transformation function
}
\description{
compile user defined parameter transformations to a
pass to a SEM
}
\examples{
syntax <- "
parameters: a, b, c, d, f, g, h # important: state all parameters which you want to use!
# This must be the first line in your statement!
a = b + c
d = a*f
g = log(h)
"
fns <- .compileTransformations(syntax)
}
