% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeSEM.R
\name{.getRawData}
\alias{.getRawData}
\title{.getRawData}
\usage{
.getRawData(lavaanModel, dataSet, estimator)
}
\arguments{
\item{lavaanModel}{model fitted with lavaan}

\item{dataSet}{user supplied data set}

\item{estimator}{which estimator is used?}
}
\value{
raw data
}
\description{
Extracts the raw data from lavaan or adapts a user supplied data set to
the structure of the lavaan data
}
\keyword{internal}
