% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-regularizedSEM.R
\name{estimates,regularizedSEM-method}
\alias{estimates,regularizedSEM-method}
\title{estimates}
\usage{
\S4method{estimates}{regularizedSEM}(object, criterion = NULL, transformations = FALSE)
}
\arguments{
\item{object}{object of class regularizedSEM}

\item{criterion}{fit index (e.g., AIC) used to select the parameters}

\item{transformations}{boolean: Should transformations be returned?}
}
\value{
returns a matrix with estimates
}
\description{
estimates
}
