% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedPenalty.R
\name{fit}
\alias{fit}
\title{fit}
\usage{
fit(mixedPenalty)
}
\arguments{
\item{mixedPenalty}{object of class mixedPenalty. This object can be created
with the mixedPenalty function. Penalties can be added with the addCappedL1, addElastiNet,
addLasso, addLsp, addMcp, and addScad functions.}
}
\value{
throws error in case of undefined penalty combinations.
}
\description{
Optimizes an object with mixed penalty. See ?mixedPenalty for more details.
}
