% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIC.R
\name{GIC}
\alias{GIC}
\title{GIC}
\usage{
GIC(regularizedSEM, k = 2)
}
\arguments{
\item{regularizedSEM}{model of class regularizedSEM}

\item{k}{numeric value to scale the degrees of freedom}
}
\value{
vector with GIC values.
}
\description{
THIS FUNCTION IS UNDER DEVELOPMENT AND SHOULD \emph{NOT} BE USED.
computes the generalized information criterion as
}
\details{
-2-log-Likelihood + k*df

where the k is a numeric value with which the
degrees of freedom (df) are multiplied. To get the equivalent
of the AIC, use k = 2. For an equivalent to the BIC, set k = log(N)

See Fan & Li (2001), p. 1355 and Zhang et al. (2010), p. 314 for more details.
\itemize{
\item Zhang, Y., Li, R., & Tsai, C.-L. (2010). Regularization Parameter
Selections via Generalized Information Criterion.
Journal of the American Statistical Association, 105(489),
312–323. https://doi.org/10.1198/jasa.2009.tm08013
}
}
