% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMNET.R
\name{GLMNET}
\alias{GLMNET}
\title{GLMNET}
\usage{
GLMNET(
  SEM,
  regularizedParameterLabels,
  lambda,
  alpha,
  adaptiveLassoWeights,
  initialHessian,
  stepSize,
  sig,
  gam,
  maxIterOut,
  maxIterIn,
  maxIterLine,
  epsOut,
  epsIn,
  useMultipleConvergenceCriteria,
  parameterChangeEps,
  regM2LLChangeEps,
  verbose = 0
)
}
\arguments{
\item{SEM}{model of class Rcpp_SEMCpp.}

\item{regularizedParameterLabels}{labels of regularized parameters}

\item{lambda}{lasso tuning parameter. Higher values = higher penalty}

\item{alpha}{0<alpha<1. Controls the weight of ridge and lasso terms. alpha = 1 is lasso, alpha = 0 ridge, in between is elastic net}

\item{adaptiveLassoWeights}{vector with weights for adaptive LASSO. Set to NULL if not using adaptive LASSO}

\item{initialHessian}{option to provide an initial Hessian to the optimizer}

\item{stepSize}{Initial stepSize of the outer iteration (theta_{k+1} = theta_k + stepSize * Stepdirection)}

\item{sig}{only relevant when lineSearch = 'GLMNET'. Controls the sigma parameter in Yuan, G.-X., Ho, C.-H., & Lin, C.-J. (2012). An improved GLMNET for l1-regularized logistic regression. The Journal of Machine Learning Research, 13, 1999–2030. https://doi.org/10.1145/2020408.2020421.}

\item{gam}{Controls the gamma parameter in Yuan, G.-X., Ho, C.-H., & Lin, C.-J. (2012). An improved GLMNET for l1-regularized logistic regression. The Journal of Machine Learning Research, 13, 1999–2030. https://doi.org/10.1145/2020408.2020421. Defaults to 0.}

\item{maxIterOut}{Maximal number of outer iterations}

\item{maxIterIn}{Maximal number of inner iterations}

\item{maxIterLine}{Maximal number of iterations for the line search procedure}

\item{epsOut}{Stopping criterion for outer iterations}

\item{epsIn}{Stopping criterion for inner iterations}

\item{useMultipleConvergenceCriteria}{if set to TRUE, GLMNET will also check the change in fit and the change in parameters. If any convergence criterion is met, the optimization stops}

\item{parameterChangeEps}{if useMultipleConvergenceCriteria: change in parameters which results in convergence}

\item{regM2LLChangeEps}{if useMultipleConvergenceCriteria: change in fit which results in convergence}

\item{verbose}{0 prints no additional information, > 0 prints GLMNET iterations}
}
\description{
Performs GLMNET (see Friedman, 2010 & Yuan, 2011) with aCV4SEM:::BFGS approximated Hessian
}
