% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{lessSEM}
\alias{lessSEM}
\title{lessSEM}
\description{
lessSEM (\strong{l}essSEM \strong{e}stimates \strong{s}parse \strong{SEMr})
is an R package which provides regularized
structural equation modeling (regularized SEM)
building on lavaan.
lessSEM is heavily inspired by the \href{https://github.com/Rjacobucci/regsem}{regsem}
package.

The objectives of lessSEM are (1) to compare exact and approximate optimization of regularized SEM
and (2) to provide optimizers for other SEM packages which can be used with an
interface similar to optim

\strong{Warning}: The package is relatively new and you may
find more stable implementations of regularized SEM in the
R packages \href{https://github.com/Rjacobucci/regsem}{regsem}
and \href{https://github.com/psyphh/lslx}{lslx}.

THE SOFTWARE IS PROVIDED 'AS IS', WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN
AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
}
\author{
Jannik Orzek \href{mailto:orzek@mpib-berlin.mpg.de}{orzek@mpib-berlin.mpg.de}
}
