% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaanHelperFunctions.R
\name{lessSEM2Lavaan}
\alias{lessSEM2Lavaan}
\title{lessSEM2Lavaan}
\usage{
lessSEM2Lavaan(regularizedSEM, lambda, theta = NULL)
}
\arguments{
\item{regularizedSEM}{object created with lessSEM}

\item{lambda}{value for tuning parameter lambda}

\item{theta}{value for tuning parameter theta}
}
\value{
lavaan model
}
\description{
Creates a lavaan model object from lessSEM (only if possible).
}
\examples{
library(lessSEM)

# Identical to regsem, lessSEM builds on the lavaan
# package for model specification. The first step
# therefore is to implement the model in lavaan.

dataset <- simulateExampleData()

lavaanSyntax <- "
f =~ l1*y1 + l2*y2 + l3*y3 + l4*y4 + l5*y5 + 
     l6*y6 + l7*y7 + l8*y8 + l9*y9 + l10*y10 + 
     l11*y11 + l12*y12 + l13*y13 + l14*y14 + l15*y15
f ~~ 1*f
"

lavaanModel <- lavaan::sem(lavaanSyntax,
                           data = dataset,
                           meanstructure = TRUE,
                           std.lv = TRUE)

# Regularization:
regularized <- lasso(lavaanModel,
                     regularized = paste0("l", 11:15), 
                     lambdas = seq(0,1,.1))

lessSEM2Lavaan(regularizedSEM = regularized, 
               lambda = 1)
}
