% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothOptimization.R
\name{newTau}
\alias{newTau}
\title{newTau}
\usage{
newTau(regularizedSEM, tau)
}
\arguments{
\item{regularizedSEM}{object fitted with approximate optimization}

\item{tau}{new tau value}
}
\description{
assign new value to parameter tau used by approximate optimization
}
