% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CV4RegularizedSEM.R
\name{plot,CV4RegularizedSEM,ANY-method}
\alias{plot,CV4RegularizedSEM,ANY-method}
\title{plot}
\usage{
\S4method{plot}{CV4RegularizedSEM,ANY}(x, what = "parameters", regularizedOnly = TRUE)
}
\arguments{
\item{x}{object of class regularizedSEM}

\item{what}{either "parameters" or "fit"}

\item{regularizedOnly}{boolean: should only regularized parameters be plotted?``}
}
\description{
plots the regularized and unregularized parameters as well as the cross-validation fits for all levels of lambda
}
