% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-regularizedSEM.R
\name{plot,regularizedSEM,ANY-method}
\alias{plot,regularizedSEM,ANY-method}
\title{plot}
\usage{
\S4method{plot}{regularizedSEM,ANY}(x, regularizedOnly = TRUE)
}
\arguments{
\item{x}{object of class regularizedSEM}

\item{regularizedOnly}{boolean: should only regularized parameters be plotted?``}
}
\description{
plots the regularized and unregularized parameters for all levels of lambda
}
