% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularizeSEM.R
\name{regularizeSEM}
\alias{regularizeSEM}
\title{regularizeSEM}
\usage{
regularizeSEM(
  lavaanModel,
  regularizedParameterLabels,
  penalty,
  lambdas,
  alphas = NULL,
  adaptiveLassoWeights = NULL,
  raw = TRUE,
  control = aCV4SEM::controlGLMNET()
)
}
\arguments{
\item{lavaanModel}{model of class lavaan}

\item{regularizedParameterLabels}{labels of regularized parameters}

\item{penalty}{which penalty should be used? Available are "ridge", "lasso", "adaptiveLasso", and "elasticNet"}

\item{lambdas}{vector with lambda values. Higher values = higher penalty}

\item{alphas}{0<alpha<1. only required for elastic net. Controls the weight of ridge and lasso terms. alpha = 1 is lasso, alpha = 0 ridge}

\item{adaptiveLassoWeights}{vector with weights for adaptive LASSO. Set to NULL if not using adaptive LASSO}

\item{raw}{controls if the internal transformations of aCV4SEM is used.}

\item{control}{option to set parameters of the GLMNET optimizer}

\item{eps}{controls the smooth approximation of non-differential penalty functions (e.g., lasso, adaptive lasso, or elastic net). Smaller values result in closer approximation, but may also cause larger issues in optimization.}
}
\description{
optimize a SEM with regularized penalty function. Uses smooth approximations for non-differentiable penalty functions.
}
