% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SEMCpp}
\alias{SEMCpp}
\title{SEMCpp class}
\description{
internal SEM representation
}
\section{Fields}{

\describe{
\item{\code{new}}{Creates a new SEMCpp.}

\item{\code{fill}}{fills the SEM with the elements from an Rcpp::List}

\item{\code{addTransformation}}{adds transforamtions to a model}

\item{\code{implied}}{Computes implied means and covariance matrix}

\item{\code{fit}}{Fits the model. Returns objective value of the fitting function}

\item{\code{getParameters}}{Returns a data frame with model parameters.}

\item{\code{getEstimator}}{returns the estimator used in the model (e.g., fiml)}

\item{\code{getParameterLabels}}{Returns a vector with unique parameter labels as used internally.}

\item{\code{getGradients}}{Returns a matrix with scores.}

\item{\code{getScores}}{Returns a matrix with scores.}

\item{\code{getHessian}}{Returns the hessian of the model. Expects the labels of the
parameters and the values of the parameters as well as a boolean indicating if
these are raw. Finally, a double (eps) controls the precision of the approximation.}

\item{\code{computeTransformations}}{compute the transformations.}

\item{\code{setTransformationGradientStepSize}}{change the step size of the gradient computation for the transformations}
}}

