% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeSEM.R
\name{SEMFromLavaan}
\alias{SEMFromLavaan}
\title{SEMFromLavaan}
\usage{
SEMFromLavaan(lavaanModel, rawData, transformVariances = TRUE)
}
\arguments{
\item{lavaanModel}{model of class lavaan}

\item{rawData}{raw data set also used for lavaan model}

\item{transformVariances}{set to TRUE to use the internal transformation of variances. This will make sure that estimates for variances can never be negative}
}
\description{
internal function. Translates an object of class lavaan to the internal model representation.
}
