% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simulateExampleData}
\alias{simulateExampleData}
\title{simulateExampleData}
\usage{
simulateExampleData(
  N = 100,
  loadings = c(rep(1, 5), rep(0.4, 5), rep(0, 5)),
  percentMissing = 0
)
}
\arguments{
\item{N}{number of persons in the data set}

\item{loadings}{loadings of the latent variable on the manifest observations}

\item{percentMissing}{percentage of missing data}
}
\value{
data set for a single-factor CFA.
}
\description{
simulate data for a simple CFA model
}
\examples{
y <- lessSEM::simulateExampleData()
}
