% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-stabSel.R
\docType{class}
\name{stabSel-class}
\alias{stabSel-class}
\title{Class for stability selection}
\description{
Class for stability selection
}
\section{Slots}{

\describe{
\item{\code{regularized}}{names of regularized parameters}

\item{\code{tuningParameters}}{data.frame with tuning parameter values}

\item{\code{stabilityPaths}}{matrix with percentage of parameters being non-zero
averaged over all subsets for each setting of the tuning parameters}

\item{\code{percentSelected}}{percentage with which a parameter was selected over all
tuning parameter settings}

\item{\code{selectedParameters}}{final selected parameters}

\item{\code{settings}}{internal}
}}

