% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CV4regularizedSEMWithCustomPenalty.R
\name{summary,cv4regularizedSEMWithCustomPenalty-method}
\alias{summary,cv4regularizedSEMWithCustomPenalty-method}
\title{summary}
\usage{
\S4method{summary}{cv4regularizedSEMWithCustomPenalty}(object)
}
\description{
summary
}
