% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-regularizedSEM.R
\name{summary,regularizedSEM-method}
\alias{summary,regularizedSEM-method}
\title{summary}
\usage{
\S4method{summary}{regularizedSEM}(object, ...)
}
\arguments{
\item{object}{object of class regularizedSEM}

\item{...}{not used}
}
\value{
No return value, just prints estimates
}
\description{
summary
}
